Code classification for PheWAS

The PheWAS classification for ICD-9-CM codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.

IcdDxToPheWAS(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile.

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default, it set to True (standard description).

Value

Three new data.tables. 1) groupedDT: Based on DxDataFile with two new columns for uniform format diagnostic codes and classified standard categories. 2) summarised_groupedDT: Summarized the dataset groupedDT and sorted by memberID. 3) Error: Potential error codes from DxUniform.

Details

Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Group diagnostic codes into PheWAS

IcdDxToPheWAS(sampleDxFile, ID, ICD, Date, "2015-10-01", FALSE)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>      Decimal  ID    ICD       Date PheCode
#>   1:   Z99.2  A2   Z992 2020-05-22    <NA>
#>   2:   Z99.2  A5   Z992 2020-01-24    <NA>
#>   3:   Z99.2  A8   Z992 2015-10-27    <NA>
#>   4:   Z99.2 A13   Z992 2020-04-26    <NA>
#>   5:   Z99.2 A13   Z992 2025-02-02    <NA>
#>  ---                                      
#> 296:     650  C2    650 2017-10-10     650
#> 297:     650  D4    650 2022-04-04     650
#> 298:  123.45  C3 123.45 2022-09-28    <NA>
#> 299:  123.45  D8 123.45 2018-06-17    <NA>
#> 300:  123.45  C0 123.45 2017-05-09    <NA>
#> 
#> $summarised_groupedDT
#>      ID PheCode firstCaseDate endCaseDate count    period
#>  1:  A0  585.32    2009-07-25  2013-12-20     2 1609 days
#>  2:  A0  585.31    2010-03-28  2012-04-05     2  739 days
#>  3:  A0  585.33    2010-10-29  2010-10-29     1    0 days
#>  4:  A1  585.32    2006-11-29  2013-04-28     3 2342 days
#>  5:  A1  585.33    2014-09-24  2014-09-24     1    0 days
#>  6:  A1  585.31    2008-06-25  2008-06-25     1    0 days
#>  7:  A2  585.32    2011-09-20  2015-01-06     3 1204 days
#>  8:  A2  585.31    2015-08-12  2015-08-12     1    0 days
#>  9:  A3  585.34    2014-02-24  2014-02-24     1    0 days
#> 10:  A3  585.32    2010-11-12  2010-11-12     1    0 days
#> 11:  A3     587    2008-07-08  2013-01-31     3 1668 days
#> 12:  A4  585.31    2006-10-20  2011-01-16     4 1549 days
#> 13:  A4  585.34    2015-03-09  2015-03-09     1    0 days
#> 14:  A5     587    2009-09-10  2009-09-10     1    0 days
#> 15:  A5  585.31    2014-10-27  2015-09-22     2  330 days
#> 16:  A5  585.33    2010-08-24  2010-08-24     1    0 days
#> 17:  A6     587    2015-07-12  2015-07-12     1    0 days
#> 18:  A6  585.31    2008-06-16  2013-01-28     3 1687 days
#> 19:  A6  585.34    2007-10-01  2007-10-01     1    0 days
#> 20:  A7  585.33    2013-12-16  2013-12-16     1    0 days
#> 21:  A7  585.31    2014-06-29  2014-08-14     2   46 days
#> 22:  A7  585.34    2007-02-01  2007-02-01     1    0 days
#> 23:  A7   585.2    2012-03-31  2012-03-31     1    0 days
#> 24:  A8  585.31    2007-11-22  2014-06-03     3 2385 days
#> 25:  A8  585.32    2014-08-20  2014-08-20     1    0 days
#> 26:  A9  585.31    2009-11-07  2013-11-09     2 1463 days
#> 27:  A9   585.2    2008-12-19  2008-12-19     1    0 days
#> 28:  A9     587    2007-03-05  2008-10-17     2  592 days
#> 29: A10  585.31    2007-11-23  2012-07-30     4 1711 days
#> 30: A10   585.3    2007-11-04  2007-11-04     1    0 days
#> 31: A11  585.31    2010-09-18  2010-10-18     2   30 days
#> 32: A11  585.32    2008-03-09  2010-02-21     2  714 days
#> 33: A11     587    2011-09-03  2011-09-03     1    0 days
#> 34: A12  585.31    2011-12-28  2015-06-29     2 1279 days
#> 35: A12  585.33    2006-05-14  2006-05-14     1    0 days
#> 36: A12   585.3    2008-12-20  2011-02-25     2  797 days
#> 37: A13  585.31    2010-09-29  2010-09-29     1    0 days
#> 38: A13  585.32    2006-04-29  2010-02-21     2 1394 days
#> 39: A14  585.31    2006-11-28  2014-12-21     3 2945 days
#> 40: A14  585.32    2013-07-26  2013-07-26     1    0 days
#> 41: A14  585.33    2011-11-07  2011-11-07     1    0 days
#> 42: A15  585.31    2007-05-25  2012-10-13     3 1968 days
#> 43: A15   585.2    2008-06-28  2008-06-28     1    0 days
#> 44: A16  585.31    2007-04-15  2013-08-18     3 2317 days
#> 45: A16     587    2011-12-29  2011-12-29     1    0 days
#> 46: A16   585.2    2014-12-05  2014-12-05     1    0 days
#> 47: A17  585.32    2007-02-19  2007-02-19     1    0 days
#> 48: A17     587    2014-07-03  2014-07-03     1    0 days
#> 49: A17  585.31    2007-07-17  2011-11-08     2 1575 days
#> 50: A17  585.33    2013-09-30  2013-09-30     1    0 days
#> 51: A18   585.2    2012-06-08  2012-06-08     1    0 days
#> 52: A18  585.31    2011-08-23  2014-03-04     2  924 days
#> 53: A18  585.32    2007-04-05  2012-02-10     2 1772 days
#> 54:  D1  401.22    2006-02-12  2006-02-12     1    0 days
#> 55:  D1     260    2011-06-18  2011-06-18     1    0 days
#> 56:  D3   295.1    2014-03-23  2014-03-23     1    0 days
#> 57:  D5   495.2    2014-11-15  2014-11-15     1    0 days
#> 58:  D5   394.1    2013-07-16  2013-07-16     1    0 days
#> 59:  D6     316    2005-10-09  2005-10-09     1    0 days
#> 60:  D6   334.1    2014-06-25  2014-06-25     1    0 days
#> 61:  D7     150    2012-07-04  2012-07-04     1    0 days
#> 62:  D7   296.1    2012-01-25  2012-01-25     1    0 days
#> 63:  D7     316    2007-05-22  2007-05-22     1    0 days
#> 64:  D8     344    2010-09-27  2010-09-27     1    0 days
#> 65:  D8   295.1    2007-01-30  2007-01-30     1    0 days
#> 66: A12     008    2006-01-12  2006-01-12     1    0 days
#> 67:  A1     008    2007-11-01  2007-11-01     1    0 days
#> 68:  A0     008    2014-11-05  2014-11-05     1    0 days
#> 69:  C1     008    2011-09-22  2011-09-22     1    0 days
#> 70:  D3     008    2014-09-03  2014-09-03     1    0 days
#> 71:  D6   755.4    2007-01-24  2007-01-24     1    0 days
#> 72:  D4   755.4    2007-09-23  2007-09-23     1    0 days
#> 73:  D7   755.4    2009-01-12  2009-01-12     1    0 days
#> 74:  D7     650    2018-04-29  2018-04-29     1    0 days
#> 75:  C2     650    2017-10-10  2020-11-25     2 1142 days
#> 76:  D4     650    2022-04-04  2022-04-04     1    0 days
#>      ID PheCode firstCaseDate endCaseDate count    period
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#>