Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.

Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme.

IcdDxToCCS(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isCCSCategoryDescription = T)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

isCCSCategoryDescription
Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10, default is True

Source

ICD-9-CM CCS (2015)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-CM CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip

Details

return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

IcdDxToCCS(sampleDxFile, ID, ICD, Date, "2015-10-01", TRUE)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date Number
#>    1:  0380  B  038.0 2007-09-12   1949
#>    2:  0380  C  038.0 2008-07-27   1950
#>    3:  0380  A  038.0 2009-02-13   1951
#>    4: 04111  C 041.11 2008-05-29    498
#>    5: 04111  B 041.11 2008-01-20    524
#>   ---                                  
#> 4006: Z9981  A Z99.81 2020-08-10   3653
#> 4007: Z9981  A Z99.81 2019-12-19   3845
#> 4008: Z9981  B Z99.81 2020-09-05   3846
#> 4009: Z9981  C Z99.81 2018-06-11   3847
#> 4010: Z9981  A Z99.81 2019-12-16   3848
#>                                 CCS_CATEGORY_DESCRIPTION
#>    1:                       Septicemia (except in labor)
#>    2:                       Septicemia (except in labor)
#>    3:                       Septicemia (except in labor)
#>    4:              Bacterial infection; unspecified site
#>    5:              Bacterial infection; unspecified site
#>   ---                                                   
#> 4006: Respiratory failure; insufficiency; arrest (adult)
#> 4007: Respiratory failure; insufficiency; arrest (adult)
#> 4008: Respiratory failure; insufficiency; arrest (adult)
#> 4009: Respiratory failure; insufficiency; arrest (adult)
#> 4010: Respiratory failure; insufficiency; arrest (adult)
#> 
#> $groupedData_Long
#>      ID                      CCS_CATEGORY_DESCRIPTION firstCaseDate endCaseDate
#>   1:  B                  Septicemia (except in labor)    2007-09-12  2020-06-16
#>   2:  C                  Septicemia (except in labor)    2008-07-27  2019-02-09
#>   3:  A                  Septicemia (except in labor)    2009-02-13  2021-05-21
#>   4:  C         Bacterial infection; unspecified site    2008-05-29  2020-06-10
#>   5:  B         Bacterial infection; unspecified site    2007-09-03  2020-03-26
#>  ---                                                                           
#> 455:  B             Cancer of kidney and renal pelvis    2019-04-14  2019-04-14
#> 456:  B           Other non-epithelial cancer of skin    2018-12-12  2018-12-12
#> 457:  A Suicide and intentional self-inflicted injury    2017-11-29  2021-06-21
#> 458:  B Suicide and intentional self-inflicted injury    2019-01-24  2019-05-27
#> 459:  C Suicide and intentional self-inflicted injury    2018-11-10  2019-04-16
#>      count    period
#>   1:     2 4661 days
#>   2:     3 3849 days
#>   3:     3 4480 days
#>   4:     5 4395 days
#>   5:    10 4588 days
#>  ---                
#> 455:     1    0 days
#> 456:     1    0 days
#> 457:     2 1300 days
#> 458:     2  123 days
#> 459:     2  157 days
#>