Get the Multi-level Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.

Multi-leve Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme. Four levels exist in the multi-level diagnosis CCS for ICD-9-CM codes, and two levels exist in the multi-level diagnosis CCS for ICD-10-CM codes

IcdDxToCCSLvl(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 2, CCSLvlLabel = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
icd 10 using date

CCSLevel
Clinical Classifications Software (CCS) multiple level:1~4, CCS for ICD-10-CM only has 1~2 multiple levels

CCSLvlLabel
Clinical Classifications Software (CCS) multiple level categories/description for icd9/10, default is True

Source

ICD-9-CM CCS (2012)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-CM CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip

Details

return Multi-leve Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

IcdDxToCCSLvl(sampleDxFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date Number
#>    1:  0380  B  038.0 2007-09-12   1949
#>    2:  0380  C  038.0 2008-07-27   1950
#>    3:  0380  A  038.0 2009-02-13   1951
#>    4: 04111  C 041.11 2008-05-29    498
#>    5: 04111  B 041.11 2008-01-20    524
#>   ---                                  
#> 4006: Z9981  A Z99.81 2020-08-10   3653
#> 4007: Z9981  A Z99.81 2019-12-19   3845
#> 4008: Z9981  B Z99.81 2020-09-05   3846
#> 4009: Z9981  C Z99.81 2018-06-11   3847
#> 4010: Z9981  A Z99.81 2019-12-16   3848
#>                                          CCS_LVL_2_LABEL
#>    1:                                Bacterial infection
#>    2:                                Bacterial infection
#>    3:                                Bacterial infection
#>    4:                                Bacterial infection
#>    5:                                Bacterial infection
#>   ---                                                   
#> 4006: Respiratory failure; insufficiency; arrest (adult)
#> 4007: Respiratory failure; insufficiency; arrest (adult)
#> 4008: Respiratory failure; insufficiency; arrest (adult)
#> 4009: Respiratory failure; insufficiency; arrest (adult)
#> 4010: Respiratory failure; insufficiency; arrest (adult)
#> 
#> $groupedData_Long
#>      ID                                                  CCS_LVL_2_LABEL
#>   1:  B                                              Bacterial infection
#>   2:  C                                              Bacterial infection
#>   3:  A                                              Bacterial infection
#>   4:  C                                                  Viral infection
#>   5:  A                                                  Viral infection
#>  ---                                                                    
#> 324:  C Screening and history of mental health and substance abuse codes
#> 325:  A Screening and history of mental health and substance abuse codes
#> 326:  A                    Suicide and intentional self-inflicted injury
#> 327:  B                    Suicide and intentional self-inflicted injury
#> 328:  C                    Suicide and intentional self-inflicted injury
#>      firstCaseDate endCaseDate count    period
#>   1:    2007-09-03  2020-06-16    12 4670 days
#>   2:    2008-05-29  2020-06-10     8 4395 days
#>   3:    2007-10-15  2021-05-21    10 4967 days
#>   4:    2007-08-07  2022-05-19    68 5399 days
#>   5:    2007-08-18  2022-05-16    67 5385 days
#>  ---                                          
#> 324:    2017-11-04  2021-09-23     5 1419 days
#> 325:    2017-06-01  2017-07-23     3   52 days
#> 326:    2017-11-29  2021-06-21     2 1300 days
#> 327:    2019-01-24  2019-05-27     2  123 days
#> 328:    2018-11-10  2019-04-16     2  157 days
#>