Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes

Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.

IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD 10 using date

comorbidMethod
Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity, type `ahrq`,`charlson`, or`elix`

Source

AHRQ

Charlson

Elixhauser

ICD-9-CM Elixhauser (2012-2015)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references

ICD-10-CM Elixhauser (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp

Details

return comorbidity meseaures based on ICD diagnosis codes

Examples

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", charlson)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date Number Comorbidity
#>    1:  0380  B  038.0 2007-09-12   1949        <NA>
#>    2:  0380  C  038.0 2008-07-27   1950        <NA>
#>    3:  0380  A  038.0 2009-02-13   1951        <NA>
#>    4: 04111  C 041.11 2008-05-29    498        <NA>
#>    5: 04111  B 041.11 2008-01-20    524        <NA>
#>   ---                                              
#> 4006: Z9981  A Z99.81 2020-08-10   3653        <NA>
#> 4007: Z9981  A Z99.81 2019-12-19   3845        <NA>
#> 4008: Z9981  B Z99.81 2020-09-05   3846        <NA>
#> 4009: Z9981  C Z99.81 2018-06-11   3847        <NA>
#> 4010: Z9981  A Z99.81 2019-12-16   3848        <NA>
#> 
#> $groupedData_Long
#>     ID Comorbidity firstCaseDate endCaseDate count    period
#>  1:  C   LiverMild    2007-06-23  2021-09-02    53 5185 days
#>  2:  A   LiverMild    2007-04-27  2022-05-16    64 5498 days
#>  3:  B   LiverMild    2007-06-05  2022-02-11    58 5365 days
#>  4:  A      Cancer    2008-09-21  2020-06-08     9 4278 days
#>  5:  C      Cancer    2007-06-19  2021-07-27    13 5152 days
#>  6:  B      Cancer    2007-07-22  2021-11-04    11 5219 days
#>  7:  A        Mets    2007-07-13  2020-04-01     5 4646 days
#>  8:  B        Mets    2009-01-08  2020-01-30     5 4039 days
#>  9:  C        Mets    2008-10-06  2020-08-01     4 4317 days
#> 10:  B          DM    2007-05-21  2022-01-23    34 5361 days
#> 11:  A          DM    2007-06-22  2022-03-07    28 5372 days
#> 12:  C          DM    2007-08-02  2021-10-28    25 5201 days
#> 13:  C        DMcx    2007-08-18  2020-12-17     9 4870 days
#> 14:  A        DMcx    2007-10-15  2021-06-30     7 5007 days
#> 15:  B        DMcx    2007-10-30  2021-05-24    12 4955 days
#> 16:  C    Dementia    2007-05-11  2007-05-11     1    0 days
#> 17:  A   Paralysis    2007-04-24  2008-01-12     5  263 days
#> 18:  B   Paralysis    2007-11-10  2009-02-01     3  449 days
#> 19:  C   Paralysis    2008-10-05  2008-12-16     2   72 days
#> 20:  A       Renal    2007-05-12  2022-02-08    40 5386 days
#> 21:  A          MI    2007-08-24  2019-05-28     3 4295 days
#> 22:  B          MI    2008-05-14  2020-08-18     2 4479 days
#> 23:  C          MI    2007-05-06  2021-06-19     3 5158 days
#> 24:  A   Pulmonary    2007-10-01  2022-02-05    16 5241 days
#> 25:  C         CHF    2007-06-12  2022-06-24    13 5491 days
#> 26:  A         CHF    2007-07-13  2021-07-24    10 5125 days
#> 27:  B         CHF    2007-07-01  2021-09-25    14 5200 days
#> 28:  B      Stroke    2008-01-21  2021-04-12     7 4830 days
#> 29:  A      Stroke    2007-07-16  2021-10-07     5 5197 days
#> 30:  C      Stroke    2007-08-29  2018-10-31     5 4081 days
#> 31:  C         PVD    2008-11-27  2021-06-19     2 4587 days
#> 32:  A         PVD    2007-08-22  2022-04-28     5 5363 days
#> 33:  B         PVD    2008-02-05  2019-05-04     4 4106 days
#> 34:  B LiverSevere    2007-07-15  2021-04-22    20 5030 days
#> 35:  A LiverSevere    2007-08-24  2021-11-23    15 5205 days
#> 36:  C LiverSevere    2007-06-01  2020-03-27    18 4683 days
#> 37:  B   Pulmonary    2007-05-21  2021-08-09    22 5194 days
#> 38:  C   Pulmonary    2007-11-07  2020-04-09    12 4537 days
#> 39:  C         PUD    2008-11-14  2018-03-29     2 3422 days
#> 40:  B         PUD    2008-01-28  2019-08-27     3 4229 days
#> 41:  C       Renal    2007-04-21  2021-09-17    34 5263 days
#> 42:  B       Renal    2007-05-28  2022-05-07    37 5458 days
#> 43:  C   Rheumatic    2008-06-07  2021-08-30     4 4832 days
#> 44:  A   Rheumatic    2008-11-06  2021-03-30     4 4527 days
#> 45:  B   Rheumatic    2009-01-31  2021-03-15     4 4426 days
#> 46:  B         HIV    2017-01-19  2017-01-19     1    0 days
#> 47:  A         HIV    2022-05-04  2022-05-04     1    0 days
#> 48:  A    Dementia    2020-08-26  2020-08-26     1    0 days
#>     ID Comorbidity firstCaseDate endCaseDate count    period
#> 

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", elix)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date Number Comorbidity
#>    1:  0380  B  038.0 2007-09-12   1949        <NA>
#>    2:  0380  C  038.0 2008-07-27   1950        <NA>
#>    3:  0380  A  038.0 2009-02-13   1951        <NA>
#>    4: 04111  C 041.11 2008-05-29    498        <NA>
#>    5: 04111  B 041.11 2008-01-20    524        <NA>
#>   ---                                              
#> 4006: Z9981  A Z99.81 2020-08-10   3653        <NA>
#> 4007: Z9981  A Z99.81 2019-12-19   3845        <NA>
#> 4008: Z9981  B Z99.81 2020-09-05   3846        <NA>
#> 4009: Z9981  C Z99.81 2018-06-11   3847        <NA>
#> 4010: Z9981  A Z99.81 2019-12-16   3848        <NA>
#> 
#> $groupedData_Long
#>     ID Comorbidity firstCaseDate endCaseDate count    period
#>  1:  C       LIVER    2007-06-01  2019-03-07    39 4297 days
#>  2:  A       LIVER    2007-04-27  2022-01-16    44 5378 days
#>  3:  B       LIVER    2007-06-05  2021-08-04    48 5174 days
#>  4:  A       TUMOR    2008-09-21  2009-04-23     3  214 days
#>  5:  C       TUMOR    2007-06-19  2018-03-29     5 3936 days
#>  6:  B       TUMOR    2007-07-22  2019-08-27     5 4419 days
#>  7:  A        METS    2007-07-13  2021-11-20     7 5244 days
#>  8:  B        METS    2009-01-08  2021-06-24     5 4550 days
#>  9:  C        METS    2008-10-06  2022-03-22    11 4915 days
#> 10:  B     HYPOTHY    2008-09-11  2022-02-11    36 4901 days
#> 11:  A     HYPOTHY    2007-08-21  2022-05-16    40 5382 days
#> 12:  C     HYPOTHY    2007-10-08  2021-09-02    31 5078 days
#> 13:  B          DM    2007-05-21  2021-05-24    33 5117 days
#> 14:  A          DM    2007-10-28  2021-06-30    17 4994 days
#> 15:  C          DM    2007-08-13  2021-10-28    24 5190 days
#> 16:  C        DMCX    2007-08-02  2021-04-25    10 5015 days
#> 17:  A        DMCX    2007-06-22  2022-02-15    12 5352 days
#> 18:  B        DMCX    2007-10-30  2020-11-04     7 4754 days
#> 19:  C    WGHTLOSS    2008-04-18  2009-03-13     5  329 days
#> 20:  A    WGHTLOSS    2007-07-24  2009-02-15     4  572 days
#> 21:  B    WGHTLOSS    2007-11-16  2008-06-21     4  218 days
#> 22:  B       LYTES    2007-08-22  2020-01-30    25 4544 days
#> 23:  C       LYTES    2007-08-14  2020-08-01    16 4736 days
#> 24:  A       LYTES    2007-07-16  2020-04-01    16 4643 days
#> 25:  A       OBESE    2007-08-06  2009-06-22     9  686 days
#> 26:  C       OBESE    2007-07-06  2009-04-26     8  660 days
#> 27:  B       OBESE    2007-06-19  2008-12-29     4  559 days
#> 28:  B     BLDLOSS    2008-12-30  2008-12-30     1    0 days
#> 29:  C     ANEMDEF    2007-06-27  2021-08-30    21 5178 days
#> 30:  A     ANEMDEF    2007-08-24  2021-03-30    23 4967 days
#> 31:  B     ANEMDEF    2007-07-30  2021-03-15    23 4977 days
#> 32:  C        COAG    2007-06-29  2020-07-01    11 4751 days
#> 33:  B        COAG    2007-12-25  2022-05-02    15 5242 days
#> 34:  A        COAG    2007-07-27  2021-10-25    13 5204 days
#> 35:  B     ALCOHOL    2008-10-21  2020-01-24     4 4112 days
#> 36:  C     ALCOHOL    2008-01-05  2018-05-20     5 3788 days
#> 37:  A        DRUG    2007-08-07  2009-05-18    11  650 days
#> 38:  B        DRUG    2007-06-17  2009-04-23    10  676 days
#> 39:  A       PSYCH    2008-11-11  2008-11-11     1    0 days
#> 40:  B       PSYCH    2007-07-08  2019-10-10     5 4477 days
#> 41:  A     ALCOHOL    2007-06-16  2022-01-05     8 5317 days
#> 42:  C        DRUG    2007-05-13  2009-03-25    13  682 days
#> 43:  B     DEPRESS    2007-05-31  2022-01-23    24 5351 days
#> 44:  C     DEPRESS    2007-08-30  2021-07-12    17 5065 days
#> 45:  A     DEPRESS    2007-07-09  2022-03-07    28 5355 days
#> 46:  A        PARA    2007-04-24  2022-04-28     8 5483 days
#> 47:  B        PARA    2007-11-10  2019-05-04     5 4193 days
#> 48:  C        PARA    2008-10-05  2021-06-19     3 4640 days
#> 49:  C       VALVE    2008-10-06  2018-06-19     4 3543 days
#> 50:  C       HTNCX    2007-10-21  2007-10-21     1    0 days
#> 51:  A       HTNCX    2008-03-03  2008-03-03     1    0 days
#> 52:  B       HTNCX    2007-10-24  2007-10-24     1    0 days
#> 53:  C         HTN    2007-02-04  2009-05-05    27  821 days
#> 54:  B         HTN    2007-06-13  2009-05-09    29  696 days
#> 55:  A         HTN    2007-08-29  2009-04-25    18  605 days
#> 56:  C    HTNWOCHF    2007-06-19  2021-01-25    10 4969 days
#> 57:  A    HTNWOCHF    2007-04-28  2020-06-11    11 4793 days
#> 58:  B    HTNWOCHF    2008-05-12  2021-02-20     9 4667 days
#> 59:  A    HRENWORF    2007-06-24  2007-06-24     1    0 days
#> 60:  B    HRENWORF    2008-01-14  2008-01-14     1    0 days
#> 61:  C    HRENWORF    2007-06-05  2007-06-05     1    0 days
#> 62:  A     HRENWRF    2008-08-17  2022-01-24    25 4908 days
#> 63:  C    PULMCIRC    2009-03-14  2020-05-10     3 4075 days
#> 64:  A    PULMCIRC    2009-02-10  2019-11-17     5 3932 days
#> 65:  B       VALVE    2008-03-14  2017-09-02     2 3459 days
#> 66:  C         CHF    2007-06-12  2020-07-01    12 4768 days
#> 67:  B         CHF    2007-07-28  2021-09-25    18 5173 days
#> 68:  A         CHF    2007-09-04  2022-02-05    15 5268 days
#> 69:  C    PERIVASC    2007-10-12  2019-04-29     5 4217 days
#> 70:  B    PERIVASC    2008-02-05  2020-09-21     7 4612 days
#> 71:  A    PERIVASC    2007-08-22  2020-02-28     3 4573 days
#> 72:  B    CHRNLUNG    2007-05-21  2022-04-12    18 5440 days
#> 73:  C    CHRNLUNG    2007-11-07  2020-09-06    12 4687 days
#> 74:  A    CHRNLUNG    2007-10-01  2009-05-10     6  587 days
#> 75:  A    RENLFAIL    2007-05-12  2020-06-08    35 4776 days
#> 76:  C    RENLFAIL    2007-04-21  2021-09-17    36 5263 days
#> 77:  B    RENLFAIL    2007-05-28  2022-03-03    37 5393 days
#> 78:  C        ARTH    2008-06-07  2008-11-15     2  161 days
#> 79:  A        ARTH    2008-11-06  2008-11-06     1    0 days
#> 80:  B        ARTH    2009-01-31  2009-01-31     1    0 days
#> 81:  A       NEURO    2007-08-18  2021-12-29    11 5247 days
#> 82:  B        AIDS    2017-01-19  2021-11-01     5 1747 days
#> 83:  A        AIDS    2019-11-25  2022-05-04     4  891 days
#> 84:  A       VALVE    2018-04-12  2018-04-12     1    0 days
#> 85:  C       NEURO    2017-07-24  2022-02-22    10 1674 days
#> 86:  B       NEURO    2017-08-05  2020-10-14     9 1166 days
#> 87:  A       LYMPH    2018-02-15  2022-06-08    19 1574 days
#> 88:  C       LYMPH    2017-09-08  2021-09-18    18 1471 days
#> 89:  B       LYMPH    2017-11-24  2022-02-03    16 1532 days
#> 90:  C        AIDS    2019-05-26  2021-08-08     4  805 days
#> 91:  B       ULCER    2019-12-06  2019-12-06     1    0 days
#> 92:  B     HRENWRF    2017-06-12  2022-05-07    31 1790 days
#> 93:  C     HRENWRF    2017-05-05  2021-09-11    22 1590 days
#> 94:  B     HTNWCHF    2019-03-28  2020-08-13     2  504 days
#> 95:  C     HTNWCHF    2019-02-02  2020-12-27     2  694 days
#> 96:  A       ULCER    2018-11-26  2020-07-11     4  593 days
#> 97:  C       ULCER    2018-10-14  2020-11-14     2  762 days
#> 98:  C     BLDLOSS    2022-06-24  2022-06-24     1    0 days
#> 99:  B    PULMCIRC    2019-03-11  2022-03-18     4 1103 days
#>     ID Comorbidity firstCaseDate endCaseDate count    period
#> 

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", ahrq)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date Number Comorbidity
#>    1:  0380  B  038.0 2007-09-12   1949        <NA>
#>    2:  0380  C  038.0 2008-07-27   1950        <NA>
#>    3:  0380  A  038.0 2009-02-13   1951        <NA>
#>    4: 04111  C 041.11 2008-05-29    498        <NA>
#>    5: 04111  B 041.11 2008-01-20    524        <NA>
#>   ---                                              
#> 4015: Z9981  A Z99.81 2020-08-10   3653        <NA>
#> 4016: Z9981  A Z99.81 2019-12-19   3845        <NA>
#> 4017: Z9981  B Z99.81 2020-09-05   3846        <NA>
#> 4018: Z9981  C Z99.81 2018-06-11   3847        <NA>
#> 4019: Z9981  A Z99.81 2019-12-16   3848        <NA>
#> 
#> $groupedData_Long
#>     ID  Comorbidity firstCaseDate endCaseDate count    period
#>  1:  C        Liver    2007-06-01  2021-09-02    67 5207 days
#>  2:  A        Liver    2007-04-27  2022-05-16    76 5498 days
#>  3:  B        Liver    2007-06-05  2022-02-11    77 5365 days
#>  4:  A        Tumor    2008-09-21  2020-06-08     9 4278 days
#>  5:  C        Tumor    2007-06-19  2021-07-27    11 5152 days
#>  6:  B        Tumor    2007-07-22  2021-11-04    12 5219 days
#>  7:  A         Mets    2007-07-13  2020-04-01     6 4646 days
#>  8:  B         Mets    2009-01-08  2020-01-30     5 4039 days
#>  9:  C         Mets    2008-10-06  2020-08-01     4 4317 days
#> 10:  B  Hypothyroid    2008-09-11  2021-02-20    13 4545 days
#> 11:  A  Hypothyroid    2007-08-21  2020-06-11    14 4678 days
#> 12:  C  Hypothyroid    2007-10-08  2021-08-23     8 5068 days
#> 13:  B           DM    2007-05-21  2022-01-23    29 5361 days
#> 14:  A           DM    2007-10-28  2022-03-07    26 5244 days
#> 15:  C           DM    2007-08-13  2021-07-12    19 5082 days
#> 16:  C         DMcx    2007-08-02  2021-10-28    15 5201 days
#> 17:  A         DMcx    2007-06-22  2021-06-30    10 5122 days
#> 18:  B         DMcx    2007-10-30  2021-05-24    18 4955 days
#> 19:  C   WeightLoss    2008-04-18  2018-06-19     6 3714 days
#> 20:  A   WeightLoss    2007-07-24  2018-04-12     5 3915 days
#> 21:  B   WeightLoss    2007-11-16  2017-09-02     5 3578 days
#> 22:  B  FluidsLytes    2007-08-22  2022-02-03    37 5279 days
#> 23:  C  FluidsLytes    2007-08-14  2021-09-18    31 5149 days
#> 24:  A  FluidsLytes    2007-07-16  2022-06-08    33 5441 days
#> 25:  A      Obesity    2007-08-06  2021-12-29    19 5259 days
#> 26:  C      Obesity    2007-07-06  2022-02-22    18 5345 days
#> 27:  B      Obesity    2007-06-19  2020-10-14    13 4866 days
#> 28:  B    BloodLoss    2008-12-30  2008-12-30     1    0 days
#> 29:  C       Anemia    2007-06-27  2019-07-11    18 4397 days
#> 30:  A       Anemia    2007-08-24  2022-01-05    20 5248 days
#> 31:  B       Anemia    2007-07-30  2020-12-17    23 4889 days
#> 32:  C Coagulopathy    2007-06-29  2020-09-06    11 4818 days
#> 33:  B Coagulopathy    2007-12-25  2022-04-12    11 5222 days
#> 34:  A Coagulopathy    2007-07-27  2009-06-10     4  684 days
#> 35:  B      Alcohol    2008-10-21  2021-11-01     5 4759 days
#> 36:  C      Alcohol    2008-01-05  2021-08-08     7 4964 days
#> 37:  A        Drugs    2007-08-07  2022-02-15    17 5306 days
#> 38:  B        Drugs    2007-06-17  2020-11-04    15 4889 days
#> 39:  A    Psychoses    2008-11-11  2020-02-28     2 4126 days
#> 40:  B    Psychoses    2007-07-08  2020-09-21     7 4824 days
#> 41:  A      Alcohol    2007-06-16  2022-02-12     9 5355 days
#> 42:  C        Drugs    2007-05-13  2021-04-25    20 5096 days
#> 43:  B   Depression    2007-05-31  2022-05-02    22 5450 days
#> 44:  C   Depression    2007-08-30  2020-07-01    14 4689 days
#> 45:  A   Depression    2007-07-09  2021-10-25    24 5222 days
#> 46:  A    Paralysis    2007-04-24  2008-01-12     5  263 days
#> 47:  B    Paralysis    2007-11-10  2009-02-01     3  449 days
#> 48:  C    Paralysis    2008-10-05  2008-12-16     2   72 days
#> 49:  C     Valvular    2008-10-06  2020-11-14     5 4422 days
#> 50:  C        HTNcx    2007-06-05  2020-12-27    10 4954 days
#> 51:  A        HTNcx    2007-04-28  2020-05-23    11 4774 days
#> 52:  B        HTNcx    2007-10-24  2022-05-07    10 5309 days
#> 53:  C          HTN    2007-02-04  2021-09-11    48 5333 days
#> 54:  B          HTN    2007-06-13  2021-10-28    57 5251 days
#> 55:  A          HTN    2007-08-29  2022-01-24    39 5262 days
#> 56:  A        Renal    2007-05-12  2020-05-23    37 4760 days
#> 57:  C         PHTN    2009-03-14  2009-03-14     1    0 days
#> 58:  A         PHTN    2009-02-10  2009-02-10     1    0 days
#> 59:  B     Valvular    2008-03-14  2019-12-06     2 4284 days
#> 60:  C          CHF    2007-06-12  2022-06-24     9 5491 days
#> 61:  B          CHF    2007-07-28  2021-09-25    11 5173 days
#> 62:  A          CHF    2007-09-04  2021-07-24     6 5072 days
#> 63:  C          PVD    2007-10-12  2021-06-19     3 4999 days
#> 64:  B          PVD    2008-02-05  2019-05-04     5 4106 days
#> 65:  A          PVD    2007-08-22  2022-04-28     5 5363 days
#> 66:  B    Pulmonary    2007-05-21  2021-08-09    21 5194 days
#> 67:  C    Pulmonary    2007-11-07  2020-04-09    12 4537 days
#> 68:  A    Pulmonary    2007-10-01  2022-02-05    15 5241 days
#> 69:  C        Renal    2007-04-21  2021-09-17    32 5263 days
#> 70:  B        Renal    2007-05-28  2022-05-07    36 5458 days
#> 71:  C    Rheumatic    2008-06-07  2021-08-30     7 4832 days
#> 72:  A    Rheumatic    2008-11-06  2021-03-30     6 4527 days
#> 73:  B    Rheumatic    2009-01-31  2021-03-15     4 4426 days
#> 74:  A   NeuroOther    2007-08-18  2021-11-20     4 5208 days
#> 75:  B          HIV    2017-01-19  2017-01-19     1    0 days
#> 76:  A          HIV    2022-05-04  2022-05-04     1    0 days
#> 77:  C    Psychoses    2017-05-09  2019-04-29     3  720 days
#> 78:  C   NeuroOther    2018-12-27  2022-03-22    10 1181 days
#> 79:  B   NeuroOther    2020-06-01  2021-06-24     5  388 days
#> 80:  B         PHTN    2017-12-10  2019-10-10     2  669 days
#> 81:  A     Valvular    2018-11-26  2020-07-11     4  593 days
#> 82:  B          PUD    2019-08-27  2019-08-27     1    0 days
#> 83:  C          PUD    2018-03-29  2018-03-29     1    0 days
#>     ID  Comorbidity firstCaseDate endCaseDate count    period
#>