Get the categories of ICD-9 and ICD-10 codes on diagnoses, the grouping rules are based on your standards.

This can be used to select the first diagnosis record based on ICD code (grepl language) in DxDataFile, return first diagnosis record based on factIcd Data

IcdDxToCustom(DxDataFile, idColName, icdColName, dateColName,
  CustomGroupingTable)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

CustomGroupingTable
Grouping rules of clustering the ICD is based on yourself! There are two column in the dataframe/datatable: "group" and "ICD"

Examples

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

groupingTable <- data.frame(group = rep("Chronic kidney disease",6),
                            ICD = c("N181","5853","5854","5855","5856","5859"),
                            stringsAsFactors = FALSE)
IcdDxToCustom(sampleDxFile, ID, ICD, Date,
              CustomGroupingTable = groupingTable)

#> $groupedDT
#>         ICD  ID       Date group
#>   1:   Z992  A2 2020-05-22  <NA>
#>   2:   Z992  A5 2020-01-24  <NA>
#>   3:   Z992  A8 2015-10-27  <NA>
#>   4:   Z992 A13 2020-04-26  <NA>
#>   5:   Z992 A13 2025-02-02  <NA>
#>  ---                            
#> 296:    650  C2 2017-10-10  <NA>
#> 297:    650  D4 2022-04-04  <NA>
#> 298: 123.45  C3 2022-09-28  <NA>
#> 299: 123.45  D8 2018-06-17  <NA>
#> 300: 123.45  C0 2017-05-09  <NA>
#> 
#> $summarised_groupedDT
#>      ID                  group firstCaseDate endCaseDate count    period
#>  1:  A0 Chronic kidney disease    2009-07-25  2013-12-20     3 1609 days
#>  2:  A1 Chronic kidney disease    2006-11-29  2014-09-24     4 2856 days
#>  3: A10 Chronic kidney disease    2007-11-04  2007-11-04     1    0 days
#>  4: A11 Chronic kidney disease    2008-03-09  2010-02-21     2  714 days
#>  5: A12 Chronic kidney disease    2006-05-14  2011-02-25     3 1748 days
#>  6: A13 Chronic kidney disease    2006-04-29  2010-02-21     2 1394 days
#>  7: A14 Chronic kidney disease    2011-11-07  2013-07-26     2  627 days
#>  8: A17 Chronic kidney disease    2007-02-19  2013-09-30     2 2415 days
#>  9: A18 Chronic kidney disease    2007-04-05  2012-02-10     2 1772 days
#> 10:  A2 Chronic kidney disease    2011-09-20  2015-01-06     3 1204 days
#> 11:  A3 Chronic kidney disease    2010-11-12  2014-02-24     2 1200 days
#> 12:  A4 Chronic kidney disease    2015-03-09  2015-03-09     1    0 days
#> 13:  A5 Chronic kidney disease    2010-08-24  2010-08-24     1    0 days
#> 14:  A6 Chronic kidney disease    2007-10-01  2007-10-01     1    0 days
#> 15:  A7 Chronic kidney disease    2007-02-01  2013-12-16     2 2510 days
#> 16:  A8 Chronic kidney disease    2014-08-20  2014-08-20     1    0 days
#>