Get the Phecode or description of ICD-9 diagnosis codes

This can be used to group Phecode or description based on ICD-9 codes in clinical diagnostic data.

IcdDxToPhecode(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
icd 10 using date

isDescription
Phecode/ description for icd9. By default it is set to True.

Source

ICD-9-Phecode (version 1.2, 2015)

https://phewascatalog.org/phecodes

Details

return Phecode or description based on ICD-9-CM codes

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

IcdDxToPhecode(sampleDxFile, ID, ICD, Date, "2015-10-01", FALSE)

#> Wrong ICD format: total 5 ICD codes (the number of occurrences is in brackets)

#> c("75.52 (1)", "755.2 (1)", "123.45 (1)", "001 (1)", "7552 (1)")

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>        ICDD ID    ICD       Date PheCode
#>   1:  553.3  C  553.3 2008-04-03   550.2
#>   2:  553.3  A  553.3 2008-05-01   550.2
#>   3: 553.20  A 553.20 2009-02-23   550.5
#>   4: 553.20  B 553.20 2007-08-29   550.5
#>   5:  553.3  A  553.3 2008-06-27   550.2
#>  ---                                    
#> 171:  75.52  D  75.52 2006-05-03    <NA>
#> 172: 785.52  A 785.52 2006-05-03  994.21
#> 173:    001  B    001 2006-05-03     008
#> 174:   7552  D   7552 2006-05-03    <NA>
#> 175:  755.2  D  755.2 2006-05-03   755.4
#> 
#> $groupedData_Long
#>     ID PheCode firstCaseDate endCaseDate count   period
#>  1:  C   550.2    2008-04-03  2008-04-03     1   0 days
#>  2:  A   550.2    2008-05-01  2008-06-27     2  57 days
#>  3:  A   550.5    2009-02-23  2009-02-23     1   0 days
#>  4:  B   550.5    2007-08-29  2007-08-29     1   0 days
#>  5:  B     483    2008-04-29  2008-04-29     1   0 days
#>  6:  A     483    2009-06-09  2009-06-09     1   0 days
#>  7:  A  427.61    2009-02-16  2009-02-16     1   0 days
#>  8:  C  427.21    2007-06-20  2008-09-11     5 449 days
#>  9:  C   427.3    2008-01-23  2008-01-23     1   0 days
#> 10:  A  427.21    2007-09-19  2007-09-19     1   0 days
#> 11:  B  427.21    2007-07-16  2009-02-25     5 590 days
#> 12:  C   427.7    2008-12-31  2008-12-31     2   0 days
#> 13:  A   427.9    2009-03-02  2009-03-02     2   0 days
#> 14:  C   427.8    2009-02-08  2009-02-08     1   0 days
#> 15:  C   427.1    2008-02-16  2008-02-16     1   0 days
#> 16:  B   427.9    2008-04-07  2008-04-07     2   0 days
#> 17:  A  427.12    2007-07-27  2008-12-27     2 519 days
#> 18:  C  427.12    2008-03-01  2008-03-01     1   0 days
#> 19:  A  585.33    2007-08-21  2007-08-28     2   7 days
#> 20:  C  585.33    2007-09-26  2007-09-26     1   0 days
#> 21:  B  585.33    2008-07-23  2008-07-23     1   0 days
#> 22:  B  585.31    2007-05-28  2009-03-05    21 647 days
#> 23:  C  585.31    2007-04-21  2009-04-21    20 731 days
#> 24:  A  585.31    2007-05-12  2009-01-17    19 616 days
#> 25:  B  585.32    2007-07-25  2008-04-28     3 278 days
#> 26:  C  585.32    2008-01-28  2009-04-06     2 434 days
#> 27:  A   585.4    2007-08-08  2007-08-08     1   0 days
#> 28:  C     587    2008-08-26  2008-08-26     1   0 days
#> 29:  A   585.3    2007-09-27  2008-11-04     6 404 days
#> 30:  A  585.32    2009-04-26  2009-04-26     1   0 days
#> 31:  C   585.3    2007-08-28  2009-05-26     4 637 days
#> 32:  B   585.3    2007-07-10  2009-03-04     2 603 days
#> 33:  C  994.21    2008-03-30  2008-09-06     2 160 days
#> 34:  C     396    2007-12-12  2007-12-12     1   0 days
#> 35:  A  994.21    2006-05-03  2008-05-03     2 731 days
#> 36:  B  994.21    2008-01-19  2008-01-19     1   0 days
#> 37:  B     008    2006-05-03  2006-05-03     1   0 days
#> 38:  D   755.4    2006-05-03  2006-05-03     1   0 days
#>     ID PheCode firstCaseDate endCaseDate count   period
#> 
#> $Error
#>       ICD count IcdVersionInFile     WrongType Suggestion
#> 1:  75.52     1            ICD 9  Wrong format           
#> 2: 123.45     1           ICD 10  Wrong format           
#> 3: A01.05     1            ICD 9 Wrong version           
#> 4:  E03.0     1            ICD 9 Wrong version           
#> 5:  A0105     1            ICD 9 Wrong version           
#> 6:  42761     1           ICD 10 Wrong version           
#> 7:  755.2     1            ICD 9  Wrong format     755.29
#> 8:    001     1            ICD 9  Wrong format       0019
#> 9:   7552     1            ICD 9  Wrong format      75529
#>