Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on procedures.

Clinical Classifications Software (CCS) for ICD-9 and ICD-10 procedure codes in clinical diagnostic data is a procedure categorization scheme.

IcdPrToCCS(PrDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isCCSCategoryDescription = TRUE)

Arguments

PrDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of PrDataFile

icdColName
A column for ICD of PrDataFile

dateColName
A column for Date of PrDataFile

icd10usingDate
ICD-10 using date

isCCSCategoryDescription
Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10. By default it is set to True.

Source

ICD-9-PCS CCS (2015)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-PCS CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_pr_icd10pcs_2019_1.zip

Details

return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

head(samplePrFile)

#>    ID   ICD       Date
#> 1:  B  5681 2008-01-14
#> 2:  A  9774 2009-01-11
#> 3:  B 44.99 2009-05-10
#> 4:  C 07.59 2009-01-21
#> 5:  B  0205 2008-07-06
#> 6:  B  8812 2007-06-27

IcdPrToCCS(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

#>   [1] "Other OR therapeutic procedures of urinary tract"                                                                                   
#>   [2] "Other non-OR therapeutic procedures; female organs"                                                                                 
#>   [3] "Other OR upper GI therapeutic procedures"                                                                                           
#>   [4] "Therapeutic endocrine procedures"                                                                                                   
#>   [5] "Other OR therapeutic nervous system procedures"                                                                                     
#>   [6] "Other diagnostic radiology and related techniques"                                                                                  
#>   [7] "Other diagnostic radiology and related techniques"                                                                                  
#>   [8] "Fracture treatment including reposition with or without fixation; lower extremity fracture or dislocation (other than hip or femur)"
#>   [9] "Other diagnostic nervous system procedures"                                                                                         
#>  [10] "Hip replacement; total and partial"                                                                                                 
#>  [11] "Radiation therapy"                                                                                                                  
#>  [12] "Other OR therapeutic procedures of urinary tract"                                                                                   
#>  [13] "Other OR lower GI therapeutic procedures"                                                                                           
#>  [14] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [15] "Other procedures; hemic and lymphatic systems"                                                                                      
#>  [16] "Nonoperative removal of foreign body"                                                                                               
#>  [17] "Other OR upper GI therapeutic procedures"                                                                                           
#>  [18] "Other OR therapeutic procedures on joints"                                                                                          
#>  [19] "Other OR therapeutic procedures on skin subcutaneous tissue fascia and breast"                                                      
#>  [20] "Repair of cystocele and rectocele; obliteration of vaginal vault"                                                                   
#>  [21] "Other OR therapeutic procedures; male genital"                                                                                      
#>  [22] "Other intraocular therapeutic procedures"                                                                                           
#>  [23] "Other OR therapeutic procedures of urinary tract"                                                                                   
#>  [24] "Other diagnostic procedures of urinary tract"                                                                                       
#>  [25] "Insertion; revision; replacement; removal of cardiac pacemaker or cardioverter/defibrillator"                                       
#>  [26] "Other therapeutic procedures on eyelids; conjunctiva; cornea"                                                                       
#>  [27] "Removal of ectopic pregnancy"                                                                                                       
#>  [28] "Repair of retina"                                                                                                                   
#>  [29] "Other non-OR gastrointestinal therapeutic procedures"                                                                               
#>  [30] "Other OR procedures on mouth and throat"                                                                                            
#>  [31] "Other OR therapeutic procedures on bone"                                                                                            
#>  [32] "Other OR heart procedures"                                                                                                          
#>  [33] "Other OR procedures on mouth and throat"                                                                                            
#>  [34] "Other OR therapeutic procedures on joints"                                                                                          
#>  [35] "Other therapeutic procedures on the ear nose and sinus"                                                                             
#>  [36] "Other OR gastrointestinal therapeutic procedures"                                                                                   
#>  [37] "Other non-OR gastrointestinal therapeutic procedures"                                                                               
#>  [38] "Ophthalmologic and otologic diagnosis and treatment"                                                                                
#>  [39] "Other therapeutic procedures on muscles and tendons"                                                                                
#>  [40] "Skin graft"                                                                                                                         
#>  [41] "Other OR procedures on vessels of head and neck"                                                                                    
#>  [42] "Other OR procedures on vessels other than head and neck"                                                                            
#>  [43] "Other diagnostic radiology and related techniques"                                                                                  
#>  [44] "Other therapeutic procedures on eyelids; conjunctiva; cornea"                                                                       
#>  [45] "Alcohol and drug rehabilitation/detoxification"                                                                                     
#>  [46] "Other respiratory therapy"                                                                                                          
#>  [47] "Other OR therapeutic procedures on musculoskeletal system"                                                                          
#>  [48] "Other non-OR gastrointestinal therapeutic procedures"                                                                               
#>  [49] "Other OR procedures on mouth and throat"                                                                                            
#>  [50] "Other intraocular therapeutic procedures"                                                                                           
#>  [51] "Forceps; vacuum; and breech delivery"                                                                                               
#>  [52] NA                                                                                                                                   
#>  [53] NA                                                                                                                                   
#>  [54] "Diagnostic procedures on eye"                                                                                                       
#>  [55] "Other OR procedures on mouth and throat"                                                                                            
#>  [56] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [57] "Procedures on the urethra"                                                                                                          
#>  [58] "Other intraocular therapeutic procedures"                                                                                           
#>  [59] "Other therapeutic procedures"                                                                                                       
#>  [60] "Other non-OR lower GI therapeutic procedures"                                                                                       
#>  [61] "Other therapeutic procedures"                                                                                                       
#>  [62] "Diagnostic endocrine procedures"                                                                                                    
#>  [63] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [64] "Other diagnostic procedures"                                                                                                        
#>  [65] "Other OR upper GI therapeutic procedures"                                                                                           
#>  [66] "Fracture treatment including reposition with or without fixation; lower extremity fracture or dislocation (other than hip or femur)"
#>  [67] "Other OR procedures on vessels of head and neck"                                                                                    
#>  [68] "Incision and excision of CNS"                                                                                                       
#>  [69] NA                                                                                                                                   
#>  [70] "Other OR therapeutic procedures of urinary tract"                                                                                   
#>  [71] NA                                                                                                                                   
#>  [72] "Non-imaging nuclear medicine probe or assay"                                                                                        
#>  [73] "Other non-OR gastrointestinal therapeutic procedures"                                                                               
#>  [74] "Diagnostic procedures on eye"                                                                                                       
#>  [75] "Other non-OR therapeutic procedures; male genital"                                                                                  
#>  [76] "Lens and cataract procedures"                                                                                                       
#>  [77] "Other operations on ovary"                                                                                                          
#>  [78] "Other OR Rx procedures on respiratory system and mediastinum"                                                                       
#>  [79] "Other OR therapeutic procedures on musculoskeletal system"                                                                          
#>  [80] "Other therapeutic procedures on muscles and tendons"                                                                                
#>  [81] "Oophorectomy; unilateral and bilateral"                                                                                             
#>  [82] NA                                                                                                                                   
#>  [83] "Destruction of lesion of retina and choroid"                                                                                        
#>  [84] "Other therapeutic procedures on eyelids; conjunctiva; cornea"                                                                       
#>  [85] "Tympanoplasty"                                                                                                                      
#>  [86] "Other vascular bypass and shunt; not heart"                                                                                         
#>  [87] "Incision and excision of CNS"                                                                                                       
#>  [88] "Coronary artery bypass graft (CABG)"                                                                                                
#>  [89] "Other OR heart procedures"                                                                                                          
#>  [90] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [91] "Coronary artery bypass graft (CABG)"                                                                                                
#>  [92] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [93] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [94] "Other diagnostic nervous system procedures"                                                                                         
#>  [95] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [96] "Other diagnostic nervous system procedures"                                                                                         
#>  [97] "Decompression peripheral nerve"                                                                                                     
#>  [98] "Other OR therapeutic nervous system procedures"                                                                                     
#>  [99] "Other OR therapeutic nervous system procedures"                                                                                     
#> [100] "Other diagnostic nervous system procedures"                                                                                         
#> [101] "Other non-OR or closed therapeutic nervous system procedures"                                                                       
#> [102] "Other OR therapeutic nervous system procedures"                                                                                     
#> [103] "Other non-OR or closed therapeutic nervous system procedures"                                                                       
#> [104] "Other diagnostic nervous system procedures"                                                                                         
#> [105] "Other OR therapeutic nervous system procedures"                                                                                     
#> [106] "Other OR therapeutic nervous system procedures"                                                                                     
#> [107] "Other OR therapeutic nervous system procedures"                                                                                     
#> [108] "Other diagnostic nervous system procedures"                                                                                         
#> [109] "Incision and excision of CNS"                                                                                                       
#> [110] "Other OR therapeutic nervous system procedures"                                                                                     
#> [111] "Other OR therapeutic nervous system procedures"                                                                                     
#> [112] "Other OR heart procedures"                                                                                                          
#> [113] "Other OR therapeutic nervous system procedures"                                                                                     
#> [114] "Decompression peripheral nerve"                                                                                                     
#> [115] "Other vascular bypass and shunt; not heart"                                                                                         
#> [116] "Other OR therapeutic nervous system procedures"                                                                                     
#> [117] "Other OR therapeutic nervous system procedures"                                                                                     
#> [118] "Other diagnostic nervous system procedures"                                                                                         
#> [119] "Other OR heart procedures"                                                                                                          
#> [120] "Radiation therapy"                                                                                                                  
#> [121] "Other OR therapeutic nervous system procedures"                                                                                     
#> [122] "Other OR therapeutic nervous system procedures"                                                                                     
#> [123] "Other OR therapeutic nervous system procedures"                                                                                     
#> [124] "Other OR therapeutic nervous system procedures"                                                                                     
#> [125] "Other vascular bypass and shunt; not heart"                                                                                         
#> [126] "Other non-OR or closed therapeutic nervous system procedures"                                                                       
#> [127] "Other OR therapeutic nervous system procedures"                                                                                     
#> [128] "Other OR therapeutic nervous system procedures"                                                                                     
#> [129] "Other OR heart procedures"                                                                                                          
#> [130] "Other OR therapeutic nervous system procedures"                                                                                     
#> [131] "Other OR therapeutic nervous system procedures"                                                                                     
#> [132] "Coronary artery bypass graft (CABG)"                                                                                                
#> [133] "Insertion of catheter or spinal stimulator and injection into spinal canal"                                                         
#> [134] "Other OR therapeutic nervous system procedures"                                                                                     
#> [135] "Coronary artery bypass graft (CABG)"                                                                                                
#> [136] "Other OR heart procedures"                                                                                                          
#> [137] "Other OR therapeutic nervous system procedures"                                                                                     
#> [138] "Other OR heart procedures"                                                                                                          
#> [139] "Other OR heart procedures"                                                                                                          
#> [140] "Other OR therapeutic nervous system procedures"                                                                                     
#> [141] "Incision and excision of CNS"                                                                                                       
#> [142] "Other OR therapeutic nervous system procedures"                                                                                     
#> [143] "Other OR therapeutic nervous system procedures"                                                                                     
#> [144] "Other OR therapeutic nervous system procedures"                                                                                     
#> [145] "Other OR therapeutic nervous system procedures"                                                                                     
#> [146] "Other OR therapeutic nervous system procedures"                                                                                     
#> [147] "Other diagnostic nervous system procedures"                                                                                         
#> [148] "Other diagnostic nervous system procedures"                                                                                         
#> [149] "Other diagnostic nervous system procedures"                                                                                         
#> [150] "Other OR therapeutic nervous system procedures"                                                                                     
#> [151] "Insertion; replacement; or removal of extracranial ventricular shunt"                                                               
#> [152] "Other OR therapeutic nervous system procedures"                                                                                     
#> [153] "Other OR heart procedures"                                                                                                          
#> [154] "Other OR therapeutic nervous system procedures"                                                                                     
#> [155] "Other OR heart procedures"                                                                                                          
#> [156] "Other OR therapeutic nervous system procedures"                                                                                     
#> [157] "Other OR therapeutic nervous system procedures"                                                                                     
#> [158] "Other non-OR or closed therapeutic nervous system procedures"                                                                       
#> [159] "Other OR therapeutic nervous system procedures"                                                                                     
#> [160] "Other OR therapeutic nervous system procedures"                                                                                     
#> [161] "Other OR therapeutic nervous system procedures"                                                                                     
#> [162] "Other OR therapeutic nervous system procedures"                                                                                     
#> [163] "Other OR therapeutic nervous system procedures"                                                                                     
#> [164] "Other vascular bypass and shunt; not heart"                                                                                         
#> [165] "Other OR therapeutic nervous system procedures"                                                                                     
#> [166] "Other OR therapeutic nervous system procedures"                                                                                     
#> [167] "Insertion; replacement; or removal of extracranial ventricular shunt"                                                               
#> [168] "Other OR heart procedures"                                                                                                          
#> [169] "Coronary artery bypass graft (CABG)"                                                                                                
#> [170] "Other OR therapeutic nervous system procedures"