Code classification for CCS

The CCS classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data on procedures

IcdPrToCCS(PrDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isDescription = TRUE)

IcdPrToCCSLvl(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, isDescription = TRUE)

Arguments

PrDataFile
A file of clinical procedure data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A MemberID column of PrDataFile

icdColName
A ICD column of PrDataFile

dateColName
A Date column of PrDataFile

icd10usingDate
ICD-10 using date

isDescription
Category or description of standard classification methods for ICD procedure codes. By default, it set to True (standard description).

CCSLevel
By default, it set to 1. CCS multiple level:1~3, CCS for ICD-10-CM only has 1~2 multiple levels.

Value

Two new data.tables. 1) groupedDT: Based on PrDataFile with two new columns for uniform format procedure codes and classified standard categories. 2) Error: Potential error codes from PrUniform.

Details

These CCS functions collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.

See also

Examples

# sample file for example

head(samplePrFile)

#>    ID   ICD       Date
#> 1:  B  5681 2008-01-14
#> 2:  A  9774 2009-01-11
#> 3:  B 44.99 2009-05-10
#> 4:  C 07.59 2009-01-21
#> 5:  B  0205 2008-07-06
#> 6:  B  8812 2007-06-27

# Group procedure codes into single level of CCS classification

IcdPrToCCS(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

#> $groupedDT
#>        Short ID     ICD       Date
#>   1:    5681  B    5681 2008-01-14
#>   2:    9774  A    9774 2009-01-11
#>   3:    4499  B   44.99 2009-05-10
#>   4:    0759  C   07.59 2009-01-21
#>   5:    0205  B    0205 2008-07-06
#>  ---                              
#> 166: 00UH4KZ  B 00UH4KZ 2019-08-04
#> 167: 001U074  A 001U074 2020-10-04
#> 168: 02160AQ  C 02160AQ 2021-12-01
#> 169: 02134KC  A 02134KC 2020-01-30
#> 170: 00WU0YZ  B 00WU0YZ 2020-06-08
#>                                                  CCS_CATEGORY_DESCRIPTION
#>   1:                     Other OR therapeutic procedures of urinary tract
#>   2:                   Other non-OR therapeutic procedures; female organs
#>   3:                             Other OR upper GI therapeutic procedures
#>   4:                                     Therapeutic endocrine procedures
#>   5:                       Other OR therapeutic nervous system procedures
#>  ---                                                                     
#> 166:                       Other OR therapeutic nervous system procedures
#> 167: Insertion; replacement; or removal of extracranial ventricular shunt
#> 168:                                            Other OR heart procedures
#> 169:                                  Coronary artery bypass graft (CABG)
#> 170:                       Other OR therapeutic nervous system procedures
#> 
#> $Error
#> NULL
#> 

# Group procedure codes into multiple levels of CCS classification

IcdPrToCCSLvl(samplePrFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

#> $groupedDT
#>        Short ID     ICD       Date
#>   1:    5681  B    5681 2008-01-14
#>   2:    9774  A    9774 2009-01-11
#>   3:    4499  B   44.99 2009-05-10
#>   4:    0759  C   07.59 2009-01-21
#>   5:    0205  B    0205 2008-07-06
#>  ---                              
#> 166: 00UH4KZ  B 00UH4KZ 2019-08-04
#> 167: 001U074  A 001U074 2020-10-04
#> 168: 02160AQ  C 02160AQ 2021-12-01
#> 169: 02134KC  A 02134KC 2020-01-30
#> 170: 00WU0YZ  B 00WU0YZ 2020-06-08
#>                                                           CCS_LVL_2_LABEL
#>   1:                     Other OR therapeutic procedures of urinary tract
#>   2:                   Other non-OR therapeutic procedures; female organs
#>   3:                             Other OR upper GI therapeutic procedures
#>   4:                               Other therapeutic endocrine procedures
#>   5:                       Other OR therapeutic nervous system procedures
#>  ---                                                                     
#> 166:                       Other OR therapeutic nervous system procedures
#> 167: Insertion; replacement; or removal of extracranial ventricular shunt
#> 168:                                            Other OR heart procedures
#> 169:                                  Coronary artery bypass graft (CABG)
#> 170:                       Other OR therapeutic nervous system procedures
#> 
#> $Error
#> NULL
#>