Common argument of diagnostic functions

Common argument of diagnostic functions

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile.

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default, it set to True (standard description).

groupDataType
Four Stratified methods can chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), precise or fuzzy customized method ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it set to "ccs".

CustomGroupingTable
Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.

selectedCaseFile
Table for selectedCases. Default is 'NULL'

Details

Common argument of diagnostic functions