Data format transformation

Convert the long format of grouped data into a wide format which is fit to others analytical and plotting packages.

groupedDataLongToWide(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupDataType = ccs, CustomGroupingTable,
  isDescription = TRUE, numericOrBinary = B, selectedCaseFile = NULL)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile.

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

groupDataType
Four Stratified methods can chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), precise or fuzzy customized method ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it set to "ccs".

CustomGroupingTable
Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default, it set to True (standard description).

numericOrBinary
Members have same diagnostic categories, type `N` or `B`, default is Binary 'B'.

selectedCaseFile
Table for selectedCases. Default is 'NULL'

Details

The output dataset of this function can be numeric or binary wide format

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

# Convert long format grouped data into a wide numeric format with selected case

groupedData_Wide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                          "2015-10-01", elix,
                                          numericOrBinary = N,
                                          isDescription = FALSE,
                                          selectedCaseFile = selectedCaseFile)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

head(groupedData_Wide)

#>     ID ELIX_ARTH ELIX_CHRNLUNG ELIX_DMCX ELIX_DRUG ELIX_HRENWORF ELIX_HTNPREG
#> 1:  A0         0             0         0         0             0            0
#> 2:  A1         0             0         0         0             0            0
#> 3: A10         0             0         0         0             0            0
#> 4: A11         0             0         0         0             0            0
#> 5: A12         0             0         0         0             0            0
#> 6: A13         0             0         0         0             0            0
#>    ELIX_LIVER ELIX_LYMPH ELIX_NEURO ELIX_OBESE ELIX_PARA ELIX_PERIVASC
#> 1:          0          0          0          0         0             0
#> 2:          0          0          0          0         0             0
#> 3:          0          0          0          0         0             0
#> 4:          0          0          0          0         0             0
#> 5:          0          0          0          0         0             0
#> 6:          0          0          0          0         0             0
#>    ELIX_PSYCH ELIX_RENLFAIL ELIX_TUMOR ELIX_ULCER ELIX_VALVE ELIX_WGHTLOSS
#> 1:          0             5          0          0          0             0
#> 2:          0             5          0          0          0             0
#> 3:          0             5          0          0          0             0
#> 4:          0             5          0          0          0             0
#> 5:          0             5          0          0          0             0
#> 6:          0             5          0          0          0             0
#>    selectedCase
#> 1:     Selected
#> 2:     Selected
#> 3:     Selected
#> 4:     Selected
#> 5:     Selected
#> 6:     Selected