Code classification for Comorbidity

The comorbidities classification (AHRQ, Charlson ,and Elixhauser Comorbidity) for ICD diagnostic codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.

icdDxToComorbid(dxDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, comorbidMethod,
  isDescription = FALSE)

Arguments

dxDataFile
A file of clinical diagnostic data with at least 3 columns: "Member ID", "ICD", and "Date"

idColName
Column name for Member ID of dxDataFile

icdColName
Column name for ICD of dxDataFile

dateColName
Column name for Date of dxDataFile

icdVerColName
Column name for ICD Version used of dxDataFile. ICD version should be recorded as "9" or "10". If this argument is provided, then icd10usingDate is not needed.

icd10usingDate
The date that ICD-10 was started using

comorbidMethod
Three comorbidity methods: AHRQ, Charlson and Elixhauser Comorbidity. Change it to any of the other possible variables (ahrq,charlson, and elix).

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default it is set to FALSE ( Comorbidity categories).

Value

Three new data.tables. 1) groupedDT: Based on dxDataFile with two new columns for uniform format diagnostic codes and classified standard categories. 2) summarised_groupedDT: Summarized the dataset groupedDT and sorted by member ID. 3) Error: Potential error codes from dxUniform.

Details

Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Group diagnostic codes into charlson comorbidity categories

icdDxToComorbid(dxDataFile = sampleDxFile, idColName = ID, icdColName = ICD, dateColName = Date, icdVerColName = NULL, icd10usingDate = "2015/10/01", comorbidMethod = charlson, isDescription = FALSE)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>       Short  ID    ICD       Date Comorbidity
#>   1:   Z992  A2   Z992 2020-05-22          RD
#>   2:   Z992  A5   Z992 2020-01-24          RD
#>   3:   Z992  A8   Z992 2015-10-27          RD
#>   4:   Z992 A13   Z992 2020-04-26          RD
#>   5:   Z992 A13   Z992 2025-02-02          RD
#>  ---                                         
#> 296:    650  C2    650 2017-10-10        <NA>
#> 297:    650  D4    650 2022-04-04        <NA>
#> 298: 123.45  C3 123.45 2022-09-28        <NA>
#> 299: 123.45  D8 123.45 2018-06-17        <NA>
#> 300: 123.45  C0 123.45 2017-05-09        <NA>
#> 
#> $summarised_groupedDT
#>      ID Comorbidity firstCaseDate endCaseDate count    period
#>  1:  A0          RD    2009-07-25  2013-12-20     5 1609 days
#>  2:  A1          RD    2006-11-29  2014-09-24     5 2856 days
#>  3: A10          RD    2007-11-04  2012-07-30     5 1730 days
#>  4: A11          RD    2008-03-09  2011-09-03     5 1273 days
#>  5: A11      DIAB_C    2015-12-16  2015-12-16     1    0 days
#>  6: A12          RD    2006-05-14  2015-06-29     5 3333 days
#>  7: A13          RD    2006-04-29  2025-02-02     5 6854 days
#>  8: A14          RD    2006-11-28  2014-12-21     5 2945 days
#>  9: A15          RD    2007-05-25  2023-05-12     5 5831 days
#> 10: A16          RD    2007-04-15  2014-12-05     5 2791 days
#> 11: A17          RD    2007-02-19  2014-07-03     5 2691 days
#> 12: A18          RD    2007-04-05  2014-03-04     5 2525 days
#> 13:  A2          RD    2011-09-20  2020-05-22     5 3167 days
#> 14:  A3          RD    2008-07-08  2014-02-24     5 2057 days
#> 15:  A4          RD    2006-10-20  2015-03-09     5 3062 days
#> 16:  A5          RD    2009-09-10  2020-01-24     5 3788 days
#> 17:  A6          RD    2007-10-01  2015-07-12     5 2841 days
#> 18:  A7          RD    2007-02-01  2014-08-14     5 2751 days
#> 19:  A8          RD    2007-11-22  2015-10-27     5 2896 days
#> 20:  A9          RD    2007-03-05  2013-11-09     5 2441 days
#> 21:  B0          RD    2015-12-26  2024-02-12     6 2970 days
#> 22:  B1          RD    2016-08-08  2024-03-04     6 2765 days
#> 23:  B2          RD    2016-03-20  2024-09-20     6 3106 days
#> 24:  B3          RD    2019-05-07  2025-05-25     6 2210 days
#> 25:  B4          RD    2015-12-02  2025-07-21     6 3519 days
#> 26:  C0      CANCER    2015-12-05  2025-02-21     5 3366 days
#> 27:  C1      CANCER    2018-02-13  2025-02-20     5 2564 days
#> 28:  C2      CANCER    2016-09-09  2024-12-12     5 3016 days
#> 29:  C2      DIAB_C    2017-07-08  2017-07-08     1    0 days
#> 30:  C3      CANCER    2015-12-06  2021-10-22     4 2147 days
#> 31:  C4      CANCER    2017-08-29  2025-09-05     5 2929 days
#> 32:  C4      DIAB_C    2015-11-23  2015-11-23     1    0 days
#> 33:  D0         PUD    2017-12-24  2017-12-24     1    0 days
#> 34:  D0      DIAB_C    2017-01-29  2020-04-04     3 1161 days
#> 35:  D1       Rheum    2016-12-06  2016-12-06     1    0 days
#> 36:  D1        MSLD    2024-04-04  2024-04-04     1    0 days
#> 37:  D2      CANCER    2019-10-13  2019-10-13     1    0 days
#> 38:  D3         PVD    2017-11-19  2017-11-19     1    0 days
#> 39:  D4      CANCER    2020-03-10  2020-03-10     1    0 days
#> 40:  D4      DIAB_C    2019-01-03  2019-01-03     1    0 days
#> 41:  D4        CEVD    2020-10-06  2020-10-06     1    0 days
#> 42:  D5      CANCER    2021-02-13  2021-02-13     1    0 days
#> 43:  D5       Rheum    2022-10-10  2022-10-10     1    0 days
#> 44:  D5        COPD    2014-11-15  2014-11-15     1    0 days
#> 45:  D5      DIAB_C    2023-07-23  2023-07-23     1    0 days
#> 46:  D7       Rheum    2021-12-03  2021-12-03     1    0 days
#> 47:  D7      CANCER    2012-07-04  2012-07-04     1    0 days
#> 48:  D7      DIAB_C    2021-08-21  2021-08-21     1    0 days
#> 49:  D8         PVD    2021-03-18  2021-03-18     1    0 days
#> 50:  D8      CANCER    2025-09-10  2025-09-10     1    0 days
#> 51:  D8        PARA    2010-09-27  2010-09-27     1    0 days
#>      ID Comorbidity firstCaseDate endCaseDate count    period
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#> 

# Group diagnostic codes into ahrq comorbidity categories

icdDxToComorbid(dxDataFile = sampleDxFile, idColName = ID, icdColName = ICD, dateColName = Date, icdVerColName = NULL, icd10usingDate = "2015/10/01", comorbidMethod = ahrq, isDescription = TRUE)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>       Short  ID    ICD       Date   Description
#>   1:   Z992  A2   Z992 2020-05-22 Renal failure
#>   2:   Z992  A5   Z992 2020-01-24 Renal failure
#>   3:   Z992  A8   Z992 2015-10-27 Renal failure
#>   4:   Z992 A13   Z992 2020-04-26 Renal failure
#>   5:   Z992 A13   Z992 2025-02-02 Renal failure
#>  ---                                           
#> 296:    650  C2    650 2017-10-10          <NA>
#> 297:    650  D4    650 2022-04-04          <NA>
#> 298: 123.45  C3 123.45 2022-09-28          <NA>
#> 299: 123.45  D8 123.45 2018-06-17          <NA>
#> 300: 123.45  C0 123.45 2017-05-09          <NA>
#> 
#> $summarised_groupedDT
#>      ID                                     Description firstCaseDate
#>  1:  A0                                   Renal failure    2009-07-25
#>  2:  A1                                   Renal failure    2006-11-29
#>  3: A10                                   Renal failure    2007-11-04
#>  4: A11                                   Renal failure    2008-03-09
#>  5: A12                                   Renal failure    2006-05-14
#>  6: A13                                   Renal failure    2006-04-29
#>  7: A14                                   Renal failure    2006-11-28
#>  8: A15                                   Renal failure    2007-05-25
#>  9: A16                                   Renal failure    2007-04-15
#> 10: A17                                   Renal failure    2007-02-19
#> 11: A18                                   Renal failure    2007-04-05
#> 12:  A2                                   Renal failure    2011-09-20
#> 13:  A3                                   Renal failure    2008-07-08
#> 14:  A4                                   Renal failure    2006-10-20
#> 15:  A5                                   Renal failure    2009-09-10
#> 16:  A6                                   Renal failure    2007-10-01
#> 17:  A7                                   Renal failure    2007-02-01
#> 18:  A8                                   Renal failure    2007-11-22
#> 19:  A9                                   Renal failure    2007-03-05
#> 20:  B0                                   Renal failure    2015-12-26
#> 21:  B1                                   Renal failure    2016-08-08
#> 22:  B2                                   Renal failure    2016-03-20
#> 23:  B3                                   Renal failure    2019-05-07
#> 24:  B4                                   Renal failure    2015-12-02
#> 25:  C0                  Solid tumor without metastasis    2015-12-05
#> 26:  C1                  Solid tumor without metastasis    2018-02-13
#> 27:  C2                  Solid tumor without metastasis    2016-09-09
#> 28:  C3                  Solid tumor without metastasis    2018-01-16
#> 29:  C4                  Solid tumor without metastasis    2017-08-29
#> 30:  D0 Rheumatoid arthritis/collagen vascular diseases    2020-02-01
#> 31:  D0         Peptic ulcer disease excluding bleeding    2017-12-24
#> 32:  D0                                       Psychoses    2022-12-01
#> 33:  D1 Rheumatoid arthritis/collagen vascular diseases    2016-12-06
#> 34:  D1                                   Liver disease    2024-04-04
#> 35:  D1                       Hypertension, complicated    2006-02-12
#> 36:  D1                                     Weight loss    2011-06-18
#> 37:  D2                    Other neurological disorders    2023-10-23
#> 38:  D2                                      Drug abuse    2022-09-15
#> 39:  D2                                        Lymphoma    2019-10-13
#> 40:  D2                                Valvular disease    2016-11-21
#> 41:  D3                    Other neurological disorders    2024-01-27
#> 42:  D3                   Peripheral vascular disorders    2017-11-19
#> 43:  D3                                         Obesity    2016-01-22
#> 44:  D3                                       Psychoses    2014-03-23
#> 45:  D4                                        Lymphoma    2020-03-10
#> 46:  D4                       Hypertension, complicated    2020-01-12
#> 47:  D4                                       Paralysis    2020-10-06
#> 48:  D5                                        Lymphoma    2021-02-13
#> 49:  D5 Rheumatoid arthritis/collagen vascular diseases    2022-10-10
#> 50:  D5                       Chronic pulmonary disease    2014-11-15
#> 51:  D5                                Valvular disease    2013-07-16
#> 52:  D6 Rheumatoid arthritis/collagen vascular diseases    2024-05-21
#> 53:  D6                                      Drug abuse    2005-10-09
#> 54:  D6                    Other neurological disorders    2014-06-25
#> 55:  D7 Rheumatoid arthritis/collagen vascular diseases    2021-12-03
#> 56:  D7                  Solid tumor without metastasis    2012-07-04
#> 57:  D7                                       Psychoses    2012-01-25
#> 58:  D7                                      Drug abuse    2007-05-22
#> 59:  D8                   Peripheral vascular disorders    2021-03-18
#> 60:  D8                                        Lymphoma    2025-09-10
#> 61:  D8                                       Paralysis    2010-09-27
#> 62:  D8                                       Psychoses    2007-01-30
#>      ID                                     Description firstCaseDate
#>     endCaseDate count    period
#>  1:  2013-12-20     5 1609 days
#>  2:  2014-09-24     5 2856 days
#>  3:  2012-07-30     5 1730 days
#>  4:  2011-09-03     5 1273 days
#>  5:  2015-06-29     5 3333 days
#>  6:  2025-02-02     5 6854 days
#>  7:  2014-12-21     5 2945 days
#>  8:  2023-05-12     5 5831 days
#>  9:  2014-12-05     5 2791 days
#> 10:  2014-07-03     5 2691 days
#> 11:  2014-03-04     5 2525 days
#> 12:  2020-05-22     5 3167 days
#> 13:  2014-02-24     5 2057 days
#> 14:  2015-03-09     5 3062 days
#> 15:  2020-01-24     5 3788 days
#> 16:  2015-07-12     5 2841 days
#> 17:  2014-08-14     5 2751 days
#> 18:  2015-10-27     5 2896 days
#> 19:  2013-11-09     5 2441 days
#> 20:  2024-02-12     6 2970 days
#> 21:  2024-03-04     6 2765 days
#> 22:  2024-09-20     6 3106 days
#> 23:  2025-05-25     6 2210 days
#> 24:  2025-07-21     6 3519 days
#> 25:  2025-02-21     6 3366 days
#> 26:  2025-02-20     6 2564 days
#> 27:  2025-09-17     6 3295 days
#> 28:  2024-02-07     5 2213 days
#> 29:  2025-09-05     6 2929 days
#> 30:  2024-12-24     2 1788 days
#> 31:  2017-12-24     1    0 days
#> 32:  2022-12-01     1    0 days
#> 33:  2016-12-06     1    0 days
#> 34:  2024-04-04     1    0 days
#> 35:  2006-02-12     1    0 days
#> 36:  2011-06-18     1    0 days
#> 37:  2023-10-23     1    0 days
#> 38:  2022-09-15     1    0 days
#> 39:  2019-10-13     1    0 days
#> 40:  2016-11-21     1    0 days
#> 41:  2024-01-27     1    0 days
#> 42:  2017-11-19     1    0 days
#> 43:  2016-01-22     1    0 days
#> 44:  2014-03-23     1    0 days
#> 45:  2020-03-10     1    0 days
#> 46:  2020-01-12     1    0 days
#> 47:  2020-10-06     1    0 days
#> 48:  2021-02-13     1    0 days
#> 49:  2022-10-10     1    0 days
#> 50:  2014-11-15     1    0 days
#> 51:  2013-07-16     1    0 days
#> 52:  2024-05-21     1    0 days
#> 53:  2005-10-09     1    0 days
#> 54:  2014-06-25     1    0 days
#> 55:  2021-12-03     1    0 days
#> 56:  2012-07-04     1    0 days
#> 57:  2012-01-25     1    0 days
#> 58:  2007-05-22     1    0 days
#> 59:  2021-03-18     1    0 days
#> 60:  2025-09-10     1    0 days
#> 61:  2010-09-27     1    0 days
#> 62:  2007-01-30     1    0 days
#>     endCaseDate count    period
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#>