Get the condition era

A Condition Era is defined as a span of time when the member is assumed to have a given condition. Condition Eras are periods of Condition Occurrence. Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data. Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.

getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, gapDate = 30, conditionSelect = CCS,
  isDescription = TRUE, CustomGroupingTable)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
Icd 10 using date

gapDate
Length of condition era, default is 30 days

conditionSelect
Four Stratified method can be chosen: CCS (or CCS levels), phecode (only for icd-9), comorbidities, grepICD or customICD , type `ccs`,`ccslvl1`,`ccslvl2`,`ccslvl3`,`ccslvl4`, `phecode`, `ahrq`, `charlson`, `elix` `customGrepIcdGroup`, or `customIcdGroup`, default is CCS

isDescription
CCS/Phecode categories or description for ICD-CM codes, default is True

CustomGroupingTable
Table is for groupedICDMethod:`grepICD` and `customGroup`

Details

return DxDataFile with new column, condition era.

Examples

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ccs, FALSE)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>      ID CCS_CATEGORY firstCaseDate endCaseDate count Era    period
#>   1:  A           10    2007-05-15  2022-04-12    54  34 5446 days
#>   2:  A          205    2007-08-16  2022-05-16    46  31 5387 days
#>   3:  B           10    2007-06-09  2021-10-10    47  30 5237 days
#>   4:  C            6    2007-08-07  2022-05-19    61  30 5399 days
#>   5:  B           98    2007-06-13  2021-10-28    57  29 5251 days
#>  ---                                                              
#> 455:  C          659    2017-05-09  2017-05-09     1   1    0 days
#> 456:  C          670    2019-06-23  2019-06-23     1   1    0 days
#> 457:  C           86    2020-01-04  2020-01-04     1   1    0 days
#> 458:  C           88    2009-03-25  2009-03-25     1   1    0 days
#> 459:  C           90    2018-04-29  2018-04-29     1   1    0 days

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ccslvl3, FALSE)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>      ID CCS_LVL_3 firstCaseDate endCaseDate count Era    period
#>   1:  A        18    2007-04-17  2014-10-01    28  12 2724 days
#>   2:  A      5.11    2007-06-16  2009-05-04    17  12  688 days
#>   3:  B       1.5    2007-06-09  2009-05-08    17  12  699 days
#>   4:  B    17.2.4    2007-05-30  2009-06-03    20  12  735 days
#>   5:  B     1.3.2    2007-04-24  2009-06-10    25  11  778 days
#>  ---                                                           
#> 398:  C     9.6.2    2009-02-11  2009-02-11     1   1    0 days
#> 399:  C     9.6.5    2007-12-15  2007-12-15     1   1    0 days
#> 400:  C     9.6.6    2008-10-17  2008-10-17     1   1    0 days
#> 401:  C     9.7.5    2007-09-02  2007-09-02     1   1    0 days
#> 402:  C     9.7.6    2007-11-19  2007-11-19     1   1    0 days

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ICD)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>       ID    ICD firstCaseDate endCaseDate count Era    period
#>    1:  A    Z23    2017-06-28  2022-04-12    36  24 1749 days
#>    2:  B    I10    2017-06-12  2021-10-28    28  20 1599 days
#>    3:  C    Z23    2017-03-31  2022-06-23    40  20 1910 days
#>    4:  B    Z23    2017-01-16  2021-10-10    29  18 1728 days
#>    5:  A    I10    2017-06-24  2022-01-24    21  16 1675 days
#>   ---                                                        
#> 1992:  C  Z95.5    2019-08-12  2019-08-12     1   1    0 days
#> 1993:  C Z98.89    2020-01-26  2020-01-26     1   1    0 days
#> 1994:  C  Z99.2    2021-09-17  2021-09-17     1   1    0 days
#> 1995:  C  Z99.3    2018-07-29  2018-07-29     1   1    0 days
#> 1996:  C Z99.81    2018-06-11  2018-06-11     1   1    0 days

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, phecode, FALSE)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>      ID PheCode firstCaseDate endCaseDate count Era   period
#>   1:  B   070.3    2007-04-24  2009-06-10    23  12 778 days
#>   2:  B    1010    2007-06-09  2009-05-08    17  12 699 days
#>   3:  A     318    2007-06-15  2009-06-09    22  11 725 days
#>   4:  C   401.1    2007-02-04  2009-05-05    27  11 821 days
#>   5:  C  585.31    2007-04-21  2009-04-21    21  11 731 days
#>  ---                                                        
#> 700:  C     850    2008-01-30  2008-02-15     2   1  16 days
#> 701:  C     851    2007-10-07  2007-10-13     2   1   6 days
#> 702:  C     853    2008-06-26  2008-06-26     1   1   0 days
#> 703:  C     860    2007-09-14  2007-09-14     1   1   0 days
#> 704:  C   964.1    2007-10-17  2007-10-17     1   1   0 days

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ahrq)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID  Comorbidity firstCaseDate endCaseDate count Era    period
#>  1:  B          HTN    2007-06-13  2021-10-28    57  29 5251 days
#>  2:  A        Liver    2007-04-27  2022-05-16    76  27 5498 days
#>  3:  C          HTN    2007-02-04  2021-09-11    48  26 5333 days
#>  4:  A          HTN    2007-08-29  2022-01-24    39  25 5262 days
#>  5:  A  FluidsLytes    2007-07-16  2022-06-08    33  24 5441 days
#>  6:  B        Liver    2007-06-05  2022-02-11    77  24 5365 days
#>  7:  C  FluidsLytes    2007-08-14  2021-09-18    31  23 5149 days
#>  8:  C        Liver    2007-06-01  2021-09-02    67  22 5207 days
#>  9:  B           DM    2007-05-21  2022-01-23    29  20 5361 days
#> 10:  B   Depression    2007-05-31  2022-05-02    22  19 5450 days
#> 11:  B  FluidsLytes    2007-08-22  2022-02-03    37  18 5279 days
#> 12:  A   Depression    2007-07-09  2021-10-25    24  17 5222 days
#> 13:  C        Drugs    2007-05-13  2021-04-25    20  17 5096 days
#> 14:  A           DM    2007-10-28  2022-03-07    26  16 5244 days
#> 15:  A      Obesity    2007-08-06  2021-12-29    19  16 5259 days
#> 16:  B        Renal    2007-05-28  2022-05-07    36  16 5458 days
#> 17:  C      Obesity    2007-07-06  2022-02-22    18  16 5345 days
#> 18:  B       Anemia    2007-07-30  2020-12-17    23  15 4889 days
#> 19:  C           DM    2007-08-13  2021-07-12    19  15 5082 days
#> 20:  C        Renal    2007-04-21  2021-09-17    32  15 5263 days
#> 21:  A        Renal    2007-05-12  2020-05-23    37  14 4760 days
#> 22:  B    Pulmonary    2007-05-21  2021-08-09    21  14 5194 days
#> 23:  B         DMcx    2007-10-30  2021-05-24    18  13 4955 days
#> 24:  A       Anemia    2007-08-24  2022-01-05    20  12 5248 days
#> 25:  A  Hypothyroid    2007-08-21  2020-06-11    14  12 4678 days
#> 26:  A    Pulmonary    2007-10-01  2022-02-05    15  12 5241 days
#> 27:  B      Obesity    2007-06-19  2020-10-14    13  12 4866 days
#> 28:  B        Tumor    2007-07-22  2021-11-04    12  12 5219 days
#> 29:  A        Drugs    2007-08-07  2022-02-15    17  11 5306 days
#> 30:  C       Anemia    2007-06-27  2019-07-11    18  11 4397 days
#> 31:  B Coagulopathy    2007-12-25  2022-04-12    11  10 5222 days
#> 32:  B        Drugs    2007-06-17  2020-11-04    15  10 4889 days
#> 33:  B        HTNcx    2007-10-24  2022-05-07    10  10 5309 days
#> 34:  C Coagulopathy    2007-06-29  2020-09-06    11  10 4818 days
#> 35:  C         DMcx    2007-08-02  2021-10-28    15  10 5201 days
#> 36:  C   Depression    2007-08-30  2020-07-01    14  10 4689 days
#> 37:  C    Pulmonary    2007-11-07  2020-04-09    12  10 4537 days
#> 38:  C        Tumor    2007-06-19  2021-07-27    11  10 5152 days
#> 39:  A        HTNcx    2007-04-28  2020-05-23    11   9 4774 days
#> 40:  B          CHF    2007-07-28  2021-09-25    11   9 5173 days
#> 41:  B  Hypothyroid    2008-09-11  2021-02-20    13   9 4545 days
#> 42:  C          CHF    2007-06-12  2022-06-24     9   9 5491 days
#> 43:  A      Alcohol    2007-06-16  2022-02-12     9   8 5355 days
#> 44:  A         DMcx    2007-06-22  2021-06-30    10   8 5122 days
#> 45:  C        HTNcx    2007-06-05  2020-12-27    10   8 4954 days
#> 46:  C  Hypothyroid    2007-10-08  2021-08-23     8   8 5068 days
#> 47:  C   NeuroOther    2018-12-27  2022-03-22    10   8 1181 days
#> 48:  A        Tumor    2008-09-21  2020-06-08     9   7 4278 days
#> 49:  B    Psychoses    2007-07-08  2020-09-21     7   7 4824 days
#> 50:  C      Alcohol    2008-01-05  2021-08-08     7   7 4964 days
#> 51:  C    Rheumatic    2008-06-07  2021-08-30     7   7 4832 days
#> 52:  A         Mets    2007-07-13  2020-04-01     6   6 4646 days
#> 53:  A    Rheumatic    2008-11-06  2021-03-30     6   6 4527 days
#> 54:  A          CHF    2007-09-04  2021-07-24     6   5 5072 days
#> 55:  A          PVD    2007-08-22  2022-04-28     5   5 5363 days
#> 56:  A   WeightLoss    2007-07-24  2018-04-12     5   5 3915 days
#> 57:  B      Alcohol    2008-10-21  2021-11-01     5   5 4759 days
#> 58:  B         Mets    2009-01-08  2020-01-30     5   5 4039 days
#> 59:  B          PVD    2008-02-05  2019-05-04     5   5 4106 days
#> 60:  A   NeuroOther    2007-08-18  2021-11-20     4   4 5208 days
#> 61:  A    Paralysis    2007-04-24  2008-01-12     5   4  263 days
#> 62:  A     Valvular    2018-11-26  2020-07-11     4   4  593 days
#> 63:  B   NeuroOther    2020-06-01  2021-06-24     5   4  388 days
#> 64:  B    Rheumatic    2009-01-31  2021-03-15     4   4 4426 days
#> 65:  B   WeightLoss    2007-11-16  2017-09-02     5   4 3578 days
#> 66:  C         Mets    2008-10-06  2020-08-01     4   4 4317 days
#> 67:  C     Valvular    2008-10-06  2020-11-14     5   4 4422 days
#> 68:  C   WeightLoss    2008-04-18  2018-06-19     6   4 3714 days
#> 69:  A Coagulopathy    2007-07-27  2009-06-10     4   3  684 days
#> 70:  C          PVD    2007-10-12  2021-06-19     3   3 4999 days
#> 71:  A    Psychoses    2008-11-11  2020-02-28     2   2 4126 days
#> 72:  B         PHTN    2017-12-10  2019-10-10     2   2  669 days
#> 73:  B    Paralysis    2007-11-10  2009-02-01     3   2  449 days
#> 74:  B     Valvular    2008-03-14  2019-12-06     2   2 4284 days
#> 75:  C    Paralysis    2008-10-05  2008-12-16     2   2   72 days
#> 76:  C    Psychoses    2017-05-09  2019-04-29     3   2  720 days
#> 77:  A          HIV    2022-05-04  2022-05-04     1   1    0 days
#> 78:  A         PHTN    2009-02-10  2009-02-10     1   1    0 days
#> 79:  B    BloodLoss    2008-12-30  2008-12-30     1   1    0 days
#> 80:  B          HIV    2017-01-19  2017-01-19     1   1    0 days
#> 81:  B          PUD    2019-08-27  2019-08-27     1   1    0 days
#> 82:  C         PHTN    2009-03-14  2009-03-14     1   1    0 days
#> 83:  C          PUD    2018-03-29  2018-03-29     1   1    0 days
#>     ID  Comorbidity firstCaseDate endCaseDate count Era    period

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, charlson)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID Comorbidity firstCaseDate endCaseDate count Era    period
#>  1:  A   LiverMild    2007-04-27  2022-05-16    64  25 5498 days
#>  2:  C   LiverMild    2007-06-23  2021-09-02    53  24 5185 days
#>  3:  B          DM    2007-05-21  2022-01-23    34  23 5361 days
#>  4:  B   LiverMild    2007-06-05  2022-02-11    58  21 5365 days
#>  5:  A          DM    2007-06-22  2022-03-07    28  17 5372 days
#>  6:  B       Renal    2007-05-28  2022-05-07    37  17 5458 days
#>  7:  C          DM    2007-08-02  2021-10-28    25  17 5201 days
#>  8:  B LiverSevere    2007-07-15  2021-04-22    20  16 5030 days
#>  9:  C LiverSevere    2007-06-01  2020-03-27    18  16 4683 days
#> 10:  C       Renal    2007-04-21  2021-09-17    34  16 5263 days
#> 11:  A       Renal    2007-05-12  2022-02-08    40  15 5386 days
#> 12:  B   Pulmonary    2007-05-21  2021-08-09    22  15 5194 days
#> 13:  A   Pulmonary    2007-10-01  2022-02-05    16  13 5241 days
#> 14:  C      Cancer    2007-06-19  2021-07-27    13  12 5152 days
#> 15:  A LiverSevere    2007-08-24  2021-11-23    15  11 5205 days
#> 16:  B         CHF    2007-07-01  2021-09-25    14  11 5200 days
#> 17:  B      Cancer    2007-07-22  2021-11-04    11  11 5219 days
#> 18:  C         CHF    2007-06-12  2022-06-24    13  11 5491 days
#> 19:  B        DMcx    2007-10-30  2021-05-24    12  10 4955 days
#> 20:  C   Pulmonary    2007-11-07  2020-04-09    12  10 4537 days
#> 21:  A         CHF    2007-07-13  2021-07-24    10   9 5125 days
#> 22:  C        DMcx    2007-08-18  2020-12-17     9   8 4870 days
#> 23:  A      Cancer    2008-09-21  2020-06-08     9   7 4278 days
#> 24:  A        DMcx    2007-10-15  2021-06-30     7   6 5007 days
#> 25:  B      Stroke    2008-01-21  2021-04-12     7   6 4830 days
#> 26:  A        Mets    2007-07-13  2020-04-01     5   5 4646 days
#> 27:  A         PVD    2007-08-22  2022-04-28     5   5 5363 days
#> 28:  A      Stroke    2007-07-16  2021-10-07     5   5 5197 days
#> 29:  B        Mets    2009-01-08  2020-01-30     5   5 4039 days
#> 30:  C      Stroke    2007-08-29  2018-10-31     5   5 4081 days
#> 31:  A   Paralysis    2007-04-24  2008-01-12     5   4  263 days
#> 32:  A   Rheumatic    2008-11-06  2021-03-30     4   4 4527 days
#> 33:  B         PVD    2008-02-05  2019-05-04     4   4 4106 days
#> 34:  B   Rheumatic    2009-01-31  2021-03-15     4   4 4426 days
#> 35:  C        Mets    2008-10-06  2020-08-01     4   4 4317 days
#> 36:  C   Rheumatic    2008-06-07  2021-08-30     4   4 4832 days
#> 37:  A          MI    2007-08-24  2019-05-28     3   3 4295 days
#> 38:  B         PUD    2008-01-28  2019-08-27     3   3 4229 days
#> 39:  C          MI    2007-05-06  2021-06-19     3   3 5158 days
#> 40:  B          MI    2008-05-14  2020-08-18     2   2 4479 days
#> 41:  B   Paralysis    2007-11-10  2009-02-01     3   2  449 days
#> 42:  C         PUD    2008-11-14  2018-03-29     2   2 3422 days
#> 43:  C         PVD    2008-11-27  2021-06-19     2   2 4587 days
#> 44:  C   Paralysis    2008-10-05  2008-12-16     2   2   72 days
#> 45:  A    Dementia    2020-08-26  2020-08-26     1   1    0 days
#> 46:  A         HIV    2022-05-04  2022-05-04     1   1    0 days
#> 47:  B         HIV    2017-01-19  2017-01-19     1   1    0 days
#> 48:  C    Dementia    2007-05-11  2007-05-11     1   1    0 days
#>     ID Comorbidity firstCaseDate endCaseDate count Era    period

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, elix)

#> wrong ICD Format: total 13 ICD codes (the number of occurrences is in brackets)

#> c("E78.0 (8)", "M79.1 (8)", "Z22.52 (7)", "A04.7 (6)", "Z98.89 (4)", "E78.4 (3)", "M50.32 (2)", "123.45 (1)", "H34.832 (1)", "M48.06 (1)")

#> 

#> wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID Comorbidity firstCaseDate endCaseDate count Era    period
#>  1:  A     HYPOTHY    2007-08-21  2022-05-16    40  24 5382 days
#>  2:  B     HRENWRF    2017-06-12  2022-05-07    31  22 1790 days
#>  3:  B     DEPRESS    2007-05-31  2022-01-23    24  21 5351 days
#>  4:  B          DM    2007-05-21  2021-05-24    33  20 5117 days
#>  5:  B     HYPOTHY    2008-09-11  2022-02-11    36  20 4901 days
#>  6:  A     DEPRESS    2007-07-09  2022-03-07    28  19 5355 days
#>  7:  A     HRENWRF    2008-08-17  2022-01-24    25  19 4908 days
#>  8:  B    RENLFAIL    2007-05-28  2022-03-03    37  18 5393 days
#>  9:  C    RENLFAIL    2007-04-21  2021-09-17    36  18 5263 days
#> 10:  C     HYPOTHY    2007-10-08  2021-09-02    31  17 5078 days
#> 11:  A       LYMPH    2018-02-15  2022-06-08    19  16 1574 days
#> 12:  C          DM    2007-08-13  2021-10-28    24  16 5190 days
#> 13:  C     HRENWRF    2017-05-05  2021-09-11    22  16 1590 days
#> 14:  A       LIVER    2007-04-27  2022-01-16    44  15 5378 days
#> 15:  C     ANEMDEF    2007-06-27  2021-08-30    21  15 5178 days
#> 16:  A     ANEMDEF    2007-08-24  2021-03-30    23  14 4967 days
#> 17:  B     ANEMDEF    2007-07-30  2021-03-15    23  14 4977 days
#> 18:  B         CHF    2007-07-28  2021-09-25    18  14 5173 days
#> 19:  B        COAG    2007-12-25  2022-05-02    15  14 5242 days
#> 20:  A    RENLFAIL    2007-05-12  2020-06-08    35  13 4776 days
#> 21:  B    CHRNLUNG    2007-05-21  2022-04-12    18  13 5440 days
#> 22:  B       LYMPH    2017-11-24  2022-02-03    16  13 1532 days
#> 23:  C     DEPRESS    2007-08-30  2021-07-12    17  13 5065 days
#> 24:  C       LYMPH    2017-09-08  2021-09-18    18  13 1471 days
#> 25:  A         CHF    2007-09-04  2022-02-05    15  11 5268 days
#> 26:  A        COAG    2007-07-27  2021-10-25    13  11 5204 days
#> 27:  A    HTNWOCHF    2007-04-28  2020-06-11    11  11 4793 days
#> 28:  B       LYTES    2007-08-22  2020-01-30    25  11 4544 days
#> 29:  C         CHF    2007-06-12  2020-07-01    12  11 4768 days
#> 30:  C         HTN    2007-02-04  2009-05-05    27  11  821 days
#> 31:  C       LYTES    2007-08-14  2020-08-01    16  11 4736 days
#> 32:  A        DMCX    2007-06-22  2022-02-15    12  10 5352 days
#> 33:  A       LYTES    2007-07-16  2020-04-01    16  10 4643 days
#> 34:  C    CHRNLUNG    2007-11-07  2020-09-06    12  10 4687 days
#> 35:  C        COAG    2007-06-29  2020-07-01    11  10 4751 days
#> 36:  C        DRUG    2007-05-13  2009-03-25    13  10  682 days
#> 37:  C       NEURO    2017-07-24  2022-02-22    10  10 1674 days
#> 38:  A          DM    2007-10-28  2021-06-30    17   9 4994 days
#> 39:  A         HTN    2007-08-29  2009-04-25    18   9  605 days
#> 40:  A       OBESE    2007-08-06  2009-06-22     9   9  686 days
#> 41:  B         HTN    2007-06-13  2009-05-09    29   9  696 days
#> 42:  B    HTNWOCHF    2008-05-12  2021-02-20     9   9 4667 days
#> 43:  B       LIVER    2007-06-05  2021-08-04    48   9 5174 days
#> 44:  C        DMCX    2007-08-02  2021-04-25    10   9 5015 days
#> 45:  C    HTNWOCHF    2007-06-19  2021-01-25    10   9 4969 days
#> 46:  C        METS    2008-10-06  2022-03-22    11   9 4915 days
#> 47:  A       NEURO    2007-08-18  2021-12-29    11   8 5247 days
#> 48:  B       NEURO    2017-08-05  2020-10-14     9   8 1166 days
#> 49:  C       LIVER    2007-06-01  2019-03-07    39   8 4297 days
#> 50:  A     ALCOHOL    2007-06-16  2022-01-05     8   7 5317 days
#> 51:  A        METS    2007-07-13  2021-11-20     7   7 5244 days
#> 52:  A        PARA    2007-04-24  2022-04-28     8   7 5483 days
#> 53:  B    PERIVASC    2008-02-05  2020-09-21     7   7 4612 days
#> 54:  A    CHRNLUNG    2007-10-01  2009-05-10     6   6  587 days
#> 55:  A        DRUG    2007-08-07  2009-05-18    11   6  650 days
#> 56:  B        DRUG    2007-06-17  2009-04-23    10   6  676 days
#> 57:  C       OBESE    2007-07-06  2009-04-26     8   6  660 days
#> 58:  A    PULMCIRC    2009-02-10  2019-11-17     5   5 3932 days
#> 59:  B        AIDS    2017-01-19  2021-11-01     5   5 1747 days
#> 60:  B        DMCX    2007-10-30  2020-11-04     7   5 4754 days
#> 61:  B        METS    2009-01-08  2021-06-24     5   5 4550 days
#> 62:  B       PSYCH    2007-07-08  2019-10-10     5   5 4477 days
#> 63:  B       TUMOR    2007-07-22  2019-08-27     5   5 4419 days
#> 64:  C       TUMOR    2007-06-19  2018-03-29     5   5 3936 days
#> 65:  A        AIDS    2019-11-25  2022-05-04     4   4  891 days
#> 66:  A       ULCER    2018-11-26  2020-07-11     4   4  593 days
#> 67:  A    WGHTLOSS    2007-07-24  2009-02-15     4   4  572 days
#> 68:  B     ALCOHOL    2008-10-21  2020-01-24     4   4 4112 days
#> 69:  B       OBESE    2007-06-19  2008-12-29     4   4  559 days
#> 70:  B        PARA    2007-11-10  2019-05-04     5   4 4193 days
#> 71:  B    PULMCIRC    2019-03-11  2022-03-18     4   4 1103 days
#> 72:  C        AIDS    2019-05-26  2021-08-08     4   4  805 days
#> 73:  C     ALCOHOL    2008-01-05  2018-05-20     5   4 3788 days
#> 74:  C    PERIVASC    2007-10-12  2019-04-29     5   4 4217 days
#> 75:  A    PERIVASC    2007-08-22  2020-02-28     3   3 4573 days
#> 76:  B    WGHTLOSS    2007-11-16  2008-06-21     4   3  218 days
#> 77:  C        PARA    2008-10-05  2021-06-19     3   3 4640 days
#> 78:  C    PULMCIRC    2009-03-14  2020-05-10     3   3 4075 days
#> 79:  C       VALVE    2008-10-06  2018-06-19     4   3 3543 days
#> 80:  C    WGHTLOSS    2008-04-18  2009-03-13     5   3  329 days
#> 81:  A       TUMOR    2008-09-21  2009-04-23     3   2  214 days
#> 82:  B     HTNWCHF    2019-03-28  2020-08-13     2   2  504 days
#> 83:  B       VALVE    2008-03-14  2017-09-02     2   2 3459 days
#> 84:  C        ARTH    2008-06-07  2008-11-15     2   2  161 days
#> 85:  C     HTNWCHF    2019-02-02  2020-12-27     2   2  694 days
#> 86:  C       ULCER    2018-10-14  2020-11-14     2   2  762 days
#> 87:  A        ARTH    2008-11-06  2008-11-06     1   1    0 days
#> 88:  A    HRENWORF    2007-06-24  2007-06-24     1   1    0 days
#> 89:  A       HTNCX    2008-03-03  2008-03-03     1   1    0 days
#> 90:  A       PSYCH    2008-11-11  2008-11-11     1   1    0 days
#> 91:  A       VALVE    2018-04-12  2018-04-12     1   1    0 days
#> 92:  B        ARTH    2009-01-31  2009-01-31     1   1    0 days
#> 93:  B     BLDLOSS    2008-12-30  2008-12-30     1   1    0 days
#> 94:  B    HRENWORF    2008-01-14  2008-01-14     1   1    0 days
#> 95:  B       HTNCX    2007-10-24  2007-10-24     1   1    0 days
#> 96:  B       ULCER    2019-12-06  2019-12-06     1   1    0 days
#> 97:  C     BLDLOSS    2022-06-24  2022-06-24     1   1    0 days
#> 98:  C    HRENWORF    2007-06-05  2007-06-05     1   1    0 days
#> 99:  C       HTNCX    2007-10-21  2007-10-21     1   1    0 days
#>     ID Comorbidity firstCaseDate endCaseDate count Era    period