Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes

Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.

groupIcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod, NumericOrBinary = B,
  groupByDate = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
Icd 10 using date

comorbidMethod
Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity, type `ahrq`,`charlson`, or`elix`

NumericOrBinary
Member have one (or more) diagnostic comorbidities, type `N` or `B`, default is `B` (Binary)

groupByDate
Default is True.

Details

return comorbidity meseaures based on ICD diagnosis codes

Examples

DxDataFile <- data.frame(ID=c("A","A","B","B"),
                         ICD=c("40201","42577","I350","K289"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10","2016-03-10")),
                         stringsAsFactors = FALSE)

groupIcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", ahrq, N, TRUE)

#>   ID CHF Valvular PHTN PVD HTN HTNcx Paralysis NeuroOther Pulmonary DM DMcx
#> 1  A   1        0    0   0   0     1         0          0         0  0    0
#> 2  B   0        1    0   0   0     0         0          0         0  0    0
#>   Hypothyroid Renal Liver PUD HIV Lymphoma Mets Tumor Rheumatic Coagulopathy
#> 1           0     0     0   0   0        0    0     0         0            0
#> 2           0     0     0   1   0        0    0     0         0            0
#>   Obesity WeightLoss FluidsLytes BloodLoss Anemia Alcohol Drugs Psychoses
#> 1       0          0           0         0      0       0     0         0
#> 2       0          0           0         0      0       0     0         0
#>   Depression
#> 1          0
#> 2          0

groupIcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", charlson, B, TRUE)

#>   ID    MI   CHF   PVD Stroke Dementia Pulmonary Rheumatic   PUD LiverMild
#> 1  A FALSE  TRUE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
#> 2  B FALSE FALSE FALSE  FALSE    FALSE     FALSE     FALSE  TRUE     FALSE
#>      DM  DMcx Paralysis Renal Cancer LiverSevere  Mets   HIV
#> 1 FALSE FALSE     FALSE FALSE  FALSE       FALSE FALSE FALSE
#> 2 FALSE FALSE     FALSE FALSE  FALSE       FALSE FALSE FALSE

groupIcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", elix, N, TRUE)

#>   ID CHF Arrhythmia Valvular PHTN PVD HTN HTNcx Paralysis NeuroOther Pulmonary
#> 1  A   0          0        0    0   0   0     0         0          0         0
#> 2  B   0          0        0    0   0   0     0         0          0         0
#>   DM DMcx Hypothyroid Renal Liver PUD HIV Lymphoma Mets Tumor Rheumatic
#> 1  0    0           0     0     0   0   0        0    0     0         0
#> 2  0    0           0     0     0   1   0        0    0     0         0
#>   Coagulopathy Obesity WeightLoss FluidsLytes BloodLoss Anemia Alcohol Drugs
#> 1            0       0          0           0         0      0       0     0
#> 2            0       0          0           0         0      0       0     0
#>   Psychoses Depression
#> 1         0          0
#> 2         0          0