Get the Phecode or description of ICD-9 diagnosis codes

This can be used to group Phecode or description based on ICD-9 codes in clinical diagnostic data.

groupIcdToPhecode(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isPhecodeDescription = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
icd 10 using date

isPhecodeDescription
Phecode/ description for icd9, default is True

Source

https://phewascatalog.org/phecodes

Details

return Phecode or description based on ICD-9-CM codes

Examples

DxDataFile <- data.frame(ID = c("A", "A", "A"),
                         ICD = c("6929", "V433", "I350"),
                         Date = as.Date(c("2013-03-31", "2013-01-29", "2016-03-10")),
                         stringsAsFactors = FALSE)
groupIcdToPhecode(DxDataFile, ID, ICD, Date, "2016-01-01", FALSE)

#> warning ICD: I350

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "wrong Format" means the ICD has wrong format

#> Warning: "warning ICD" means the ICD classify to wrong ICD version (phecode does not have icd10) 

#> [1] "939"   "395.6" NA