Code format transformation

Converts the ICD procedure codes to a uniform format

IcdPrDecimalToShort(PrDataFile, icdColName, dateColName, icd10usingDate)

IcdPrShortToDecimal(PrDataFile, icdColName, dateColName, icd10usingDate)

Arguments

PrDataFile
A file of clinical procedure data with at least 2 columns: "ICD" and "Date"

icdColName
A ICD column of PrDataFile

dateColName
A Date column of PrDataFile

icd10usingDate
ICD-10 using date

Value

Two new data.tables. 1) ICD: Uniform format procedure codes. 2) Error: Potential error codes.

Details

These transformative function does not only convert the ICD to uniform format code but also check the potential coding error of the ICD codes’ format or version.

Examples

# sample file for example

head(samplePrFile)

#>    ID   ICD       Date
#> 1:  B  5681 2008-01-14
#> 2:  A  9774 2009-01-11
#> 3:  B 44.99 2009-05-10
#> 4:  C 07.59 2009-01-21
#> 5:  B  0205 2008-07-06
#> 6:  B  8812 2007-06-27

# convert the procedure codes to the short format

IcdPrDecimalToShort(samplePrFile,ICD,Date,"2015/10/01")

#> $ICD
#>          ICD
#>   1:    5681
#>   2:    9774
#>   3:    4499
#>   4:    0759
#>   5:    0205
#>  ---        
#> 166: 00UH4KZ
#> 167: 001U074
#> 168: 02160AQ
#> 169: 02134KC
#> 170: 00WU0YZ
#> 

# convert the procedure codes to the decimal format

IcdPrShortToDecimal(samplePrFile,ICD,Date,"2015/10/01")

#> $ICD
#>          ICD
#>   1:   56.81
#>   2:   97.74
#>   3:   44.99
#>   4:   07.59
#>   5:   02.05
#>  ---        
#> 166: 00UH4KZ
#> 167: 001U074
#> 168: 02160AQ
#> 169: 02134KC
#> 170: 00WU0YZ
#>