Data integration for patient record period

Function for researchers used for finding the first and last clinical event for a given patient as index date.

patientRecordDate(DxDataFile, idColName, icdColName, dateColName)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile.

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

Value

A new data.table based on DxDataFile with the earliest and latest admission date for each patient.

Details

The function queries the earliest and latest admission date for each patient.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Earliest and latest admission date for each patient

record <- patientRecordDate(sampleDxFile, ID, ICD, Date)
head(record)

#>     ID firstRecordDate endRecordDate
#> 1:  D6      2005-10-09    2025-01-05
#> 2: A12      2006-01-12    2022-06-12
#> 3:  D1      2006-02-12    2024-04-04
#> 4: A13      2006-04-29    2025-02-02
#> 5:  A9      2006-06-30    2023-12-10
#> 6:  D2      2006-09-01    2025-08-11