Data integration for case selection

This query function can select the cases matching defined conditions for analyses.

selectCases(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  groupDataType = ccs, CustomGroupingTable, isDescription = TRUE,
  caseCondition, caseCount, PeriodRange = c(30, 365),
  CaseName = "Selected")

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile.

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

groupDataType
Four Stratified methods can chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), precise or fuzzy customized method ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it set to "ccs".

CustomGroupingTable
Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default, it set to True (standard description).

caseCondition
Certain diseases of standard groups.

caseCount
Minimum number of diagnoses.

PeriodRange
Determine days of interest for performing the case selection. By default, it set from 30 to 365 days.

CaseName
Name of selected cases. By default, it set to "selected".

Value

A new data.table based on standard classification dataset with a new column: selectedCase

Details

User can select cases by diagnostic categories, such as CCS category, ICD codes, etc. The function also provides the options to set the minimum number of diagnoses within a specific duration. The output dataset can be passed to `groupedDataLongToWide` to create tables in wide format for statistical analytic usage.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

#select case with "Diseases of the urinary system" by level 2 of CCS classification

selectCases(DxDataFile = sampleDxFile,
            ID, ICD, Date,
            groupDataType = ccslvl2,
            icd10usingDate = "2015/10/01",
            caseCondition = "Diseases of the urinary system",
            caseCount = 1)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>      ID selectedCase count firstCaseDate endCaseDate    period MostCommonICD
#>  1:  A3     Selected     5    2008-07-08  2014-02-24 2057 days          V420
#>  2: A13     Selected     5    2006-04-29  2025-02-02 6854 days          5855
#>  3: A15     Selected     5    2007-05-25  2023-05-12 5831 days         V5631
#>  4: A18     Selected     5    2007-04-05  2014-03-04 2525 days          5855
#>  5:  A2     Selected     5    2011-09-20  2020-05-22 3167 days          5855
#>  6:  A9     Selected     5    2007-03-05  2013-11-09 2441 days          V420
#>  7:  B0     Selected     6    2015-12-26  2024-02-12 2970 days          N185
#>  8:  B1     Selected     6    2016-08-08  2024-03-04 2765 days          N183
#>  9:  B3     Selected     6    2019-05-07  2025-05-25 2210 days          N189
#> 10:  B4     Selected     6    2015-12-02  2025-07-21 3519 days          N185
#> 11:  A0     Selected     5    2009-07-25  2013-12-20 1609 days          5856
#> 12:  A1     Selected     5    2006-11-29  2014-09-24 2856 days          5856
#> 13: A10     Selected     5    2007-11-04  2012-07-30 1730 days          5859
#> 14: A11     Selected     5    2008-03-09  2011-09-03 1273 days          5855
#> 15: A12     Selected     5    2006-05-14  2015-06-29 3333 days          5853
#> 16: A14     Selected     5    2006-11-28  2014-12-21 2945 days          V560
#> 17: A16     Selected     5    2007-04-15  2014-12-05 2791 days         V5631
#> 18: A17     Selected     5    2007-02-19  2014-07-03 2691 days          5856
#> 19:  A4     Selected     5    2006-10-20  2015-03-09 3062 days         V5631
#> 20:  A5     Selected     5    2009-09-10  2020-01-24 3788 days          V420
#> 21:  A6     Selected     5    2007-10-01  2015-07-12 2841 days          5854
#> 22:  A7     Selected     5    2007-02-01  2014-08-14 2751 days          5854
#> 23:  A8     Selected     5    2007-11-22  2015-10-27 2896 days          V561
#> 24:  B2     Selected     6    2016-03-20  2024-09-20 3106 days          N189
#> 25:  C0 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 26:  C1 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 27:  C2 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 28:  C3 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 29:  C4 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 30:  D0 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 31:  D1 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 32:  D2 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 33:  D3 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 34:  D4 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 35:  D5 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 36:  D6 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 37:  D7 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#> 38:  D8 non-Selected    NA          <NA>        <NA>   NA days          <NA>
#>      ID selectedCase count firstCaseDate endCaseDate    period MostCommonICD
#>     MostCommonICDCount
#>  1:                  3
#>  2:                  2
#>  3:                  2
#>  4:                  2
#>  5:                  2
#>  6:                  2
#>  7:                  2
#>  8:                  2
#>  9:                  2
#> 10:                  2
#> 11:                  1
#> 12:                  1
#> 13:                  1
#> 14:                  1
#> 15:                  1
#> 16:                  1
#> 17:                  1
#> 18:                  1
#> 19:                  1
#> 20:                  1
#> 21:                  1
#> 22:                  1
#> 23:                  1
#> 24:                  1
#> 25:                 NA
#> 26:                 NA
#> 27:                 NA
#> 28:                 NA
#> 29:                 NA
#> 30:                 NA
#> 31:                 NA
#> 32:                 NA
#> 33:                 NA
#> 34:                 NA
#> 35:                 NA
#> 36:                 NA
#> 37:                 NA
#> 38:                 NA
#>     MostCommonICDCount