Select cases based on ICD code and the number of ICD codes

This can be used to select qualified cases from factIcd data based on the ICD code searching criteria and number of ICD code per patients in the inout factIcd dataset. Return qualified Members' data

selectCases(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  groupDataType = ICD, CustomGroupingTable, isDescription = TRUE,
  caseCondition, caseCount, INRofDayRange = c(30, 365),
  selectCaseType = "Selected")

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
Icd 10 using date

groupDataType
Four Stratified methods can be chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), grepICD or customICD ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it is set to "ccs".

CustomGroupingTable
Table is for groupDataType

isDescription
CCS/PheWAS categories or description for ICD-CM codes, default is 'TRUE'.

caseCondition
ICD selection rules with grepl expression

caseCount
a threshold of number of ICD for case selection

INRofDayRange
Determines the interval of days of interest for performing the case selection. By default it is set from 30 to 365 days.

selectCaseType
Aggregation of selected cases name. By default it is set to "selected".

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

selectCases(DxDataFile = sampleDxFile,
            ID, ICD, Date,
            groupDataType = ccslvl2,
            icd10usingDate = "2015/10/01",
            caseCondition = "Diseases of the heart",
            caseCount = 2)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>    ID selectedCase count firstCaseDate endCaseDate period MostCommonICD
#> 1:  A     Selected     4    2021-08-02  2022-02-20    202        I48.91
#> 2:  B     Selected     4    2007-07-16  2009-02-25    348        427.31
#> 3:  C     Selected     6    2007-06-20  2009-02-08    255        427.31
#> 4:  D  nonSelected    NA          <NA>        <NA>     NA          <NA>
#>    MostCommonICDCount
#> 1:                  4
#> 2:                  3
#> 3:                  3
#> 4:                 NA