Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.

Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme.

IcdDxToCCS(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isDescription = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

isDescription
Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10. By default it is set to True.

Source

ICD-9-CM CCS (2015)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-CM CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip

Details

return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

IcdDxToCCS(sampleDxFile, ID, ICD, Date, "2015-10-01", TRUE)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>       Short  ID    ICD       Date CCS_CATEGORY_DESCRIPTION
#>   1:   Z992  A2   Z992 2020-05-22   Chronic kidney disease
#>   2:   Z992  A5   Z992 2020-01-24   Chronic kidney disease
#>   3:   Z992  A8   Z992 2015-10-27   Chronic kidney disease
#>   4:   Z992 A13   Z992 2020-04-26   Chronic kidney disease
#>   5:   Z992 A13   Z992 2025-02-02   Chronic kidney disease
#>  ---                                                      
#> 296:    650  C2    650 2017-10-10                     <NA>
#> 297:    650  D4    650 2022-04-04                     <NA>
#> 298: 123.45  C3 123.45 2022-09-28                     <NA>
#> 299: 123.45  D8 123.45 2018-06-17                     <NA>
#> 300: 123.45  C0 123.45 2017-05-09                     <NA>
#> 
#> $summarised_groupedDT
#>      ID                                           CCS_CATEGORY_DESCRIPTION
#>  1:  A0                                             Chronic kidney disease
#>  2:  A1                                             Chronic kidney disease
#>  3: A10                                             Chronic kidney disease
#>  4: A11                                             Chronic kidney disease
#>  5: A12                                             Chronic kidney disease
#>  6: A13                                             Chronic kidney disease
#>  7: A14                                             Chronic kidney disease
#>  8: A15                                             Chronic kidney disease
#>  9: A15                                Acute and unspecified renal failure
#> 10: A16                                             Chronic kidney disease
#> 11: A16                                Acute and unspecified renal failure
#> 12: A17                                             Chronic kidney disease
#> 13: A18                                Acute and unspecified renal failure
#> 14: A18                                             Chronic kidney disease
#> 15:  A2                                             Chronic kidney disease
#> 16:  A3                                             Chronic kidney disease
#> 17:  A4                                             Chronic kidney disease
#> 18:  A5                                             Chronic kidney disease
#> 19:  A6                                             Chronic kidney disease
#> 20:  A7                                             Chronic kidney disease
#> 21:  A7                                Acute and unspecified renal failure
#> 22:  A8                                             Chronic kidney disease
#> 23:  A9                                             Chronic kidney disease
#> 24:  A9                                Acute and unspecified renal failure
#> 25:  B0                                             Chronic kidney disease
#> 26:  B0                                Acute and unspecified renal failure
#> 27:  B1                                Acute and unspecified renal failure
#> 28:  B1                                             Chronic kidney disease
#> 29:  B2                                             Chronic kidney disease
#> 30:  B2                                Acute and unspecified renal failure
#> 31:  B3                                             Chronic kidney disease
#> 32:  B4                                             Chronic kidney disease
#> 33:  C0                                                  Melanomas of skin
#> 34:  C0                                                  Cancer of bladder
#> 35:  C0                                 Cancer of brain and nervous system
#> 36:  C0                                            Cancer of head and neck
#> 37:  C1                                 Cancer of brain and nervous system
#> 38:  C1                                            Cancer of head and neck
#> 39:  C1                                                   Cancer of breast
#> 40:  C1                                           Cancer of bronchus; lung
#> 41:  C1                                Other non-epithelial cancer of skin
#> 42:  C2                              Cancer; other and unspecified primary
#> 43:  C2                                           Cancer of bronchus; lung
#> 44:  C2                   Malignant neoplasm without specification of site
#> 45:  C2                                            Cancer of head and neck
#> 46:  C2                                                  Anxiety disorders
#> 47:  C3                                                  Melanomas of skin
#> 48:  C3                                                    Cancer of colon
#> 49:  C3                                     Cancer of other urinary organs
#> 50:  C3                                Other non-epithelial cancer of skin
#> 51:  C3                                                   Cancer of breast
#> 52:  C3                                          Cancer of rectum and anus
#> 53:  C4                                                   Cancer of cervix
#> 54:  C4                                                  Cancer of bladder
#> 55:  C4                                                   Cancer of testis
#> 56:  C4                                Other non-epithelial cancer of skin
#> 57:  C4                                                   Cancer of breast
#> 58:  C4                                            Cancer of head and neck
#> 59:  C4                                                  Anxiety disorders
#> 60:  D0                           Rheumatoid arthritis and related disease
#> 61:  D0                                        Gastrointestinal hemorrhage
#> 62:  D0                                                     Mood disorders
#> 63:  D0    Spondylosis; intervertebral disc disorders; other back problems
#> 64:  D0                                                  Anxiety disorders
#> 65:  D1                           Rheumatoid arthritis and related disease
#> 66:  D1                                               Other liver diseases
#> 67:  D1         Hypertension with complications and secondary hypertension
#> 68:  D1                                           Nutritional deficiencies
#> 69:  D2                                   Other complications of pregnancy
#> 70:  D2                                        Substance-related disorders
#> 71:  D2                                             Non-Hodgkin`s lymphoma
#> 72:  D2                                              Heart valve disorders
#> 73:  D3                                              Epilepsy; convulsions
#> 74:  D3                           Complication of device; implant or graft
#> 75:  D3              Other nutritional; endocrine; and metabolic disorders
#> 76:  D3                        Schizophrenia and other psychotic disorders
#> 77:  D3                                                  Anxiety disorders
#> 78:  D4                                             Non-Hodgkin`s lymphoma
#> 79:  D4                               Diabetes mellitus with complications
#> 80:  D4 Hypertension complicating pregnancy; childbirth and the puerperium
#> 81:  D4                            Late effects of cerebrovascular disease
#> 82:  D5                                             Non-Hodgkin`s lymphoma
#> 83:  D5                           Rheumatoid arthritis and related disease
#> 84:  D5                                                             Asthma
#> 85:  D5                                              Heart valve disorders
#> 86:  D6                               Diabetes mellitus with complications
#> 87:  D6                           Rheumatoid arthritis and related disease
#> 88:  D6                                        Substance-related disorders
#> 89:  D6        Other hereditary and degenerative nervous system conditions
#> 90:  D7                           Rheumatoid arthritis and related disease
#> 91:  D7                                                Cancer of esophagus
#> 92:  D7                                                     Mood disorders
#> 93:  D7                                        Substance-related disorders
#> 94:  D8                           Complication of device; implant or graft
#> 95:  D8                                             Non-Hodgkin`s lymphoma
#> 96:  D8                                                          Paralysis
#> 97:  D8                        Schizophrenia and other psychotic disorders
#>      ID                                           CCS_CATEGORY_DESCRIPTION
#>     firstCaseDate endCaseDate count    period
#>  1:    2009-07-25  2013-12-20     5 1609 days
#>  2:    2006-11-29  2014-09-24     5 2856 days
#>  3:    2007-11-04  2012-07-30     5 1730 days
#>  4:    2008-03-09  2011-09-03     5 1273 days
#>  5:    2006-05-14  2015-06-29     5 3333 days
#>  6:    2006-04-29  2025-02-02     5 6854 days
#>  7:    2006-11-28  2014-12-21     5 2945 days
#>  8:    2007-05-25  2023-05-12     4 5831 days
#>  9:    2008-06-28  2008-06-28     1    0 days
#> 10:    2007-04-15  2013-08-18     4 2317 days
#> 11:    2014-12-05  2014-12-05     1    0 days
#> 12:    2007-02-19  2014-07-03     5 2691 days
#> 13:    2012-06-08  2012-06-08     1    0 days
#> 14:    2007-04-05  2014-03-04     4 2525 days
#> 15:    2011-09-20  2020-05-22     5 3167 days
#> 16:    2008-07-08  2014-02-24     5 2057 days
#> 17:    2006-10-20  2015-03-09     5 3062 days
#> 18:    2009-09-10  2020-01-24     5 3788 days
#> 19:    2007-10-01  2015-07-12     5 2841 days
#> 20:    2007-02-01  2014-08-14     4 2751 days
#> 21:    2012-03-31  2012-03-31     1    0 days
#> 22:    2007-11-22  2015-10-27     5 2896 days
#> 23:    2007-03-05  2013-11-09     4 2441 days
#> 24:    2008-12-19  2008-12-19     1    0 days
#> 25:    2015-12-26  2023-08-28     4 2802 days
#> 26:    2023-11-18  2024-02-12     2   86 days
#> 27:    2016-09-07  2016-09-07     1    0 days
#> 28:    2016-08-08  2024-03-04     5 2765 days
#> 29:    2016-03-20  2024-09-20     5 3106 days
#> 30:    2024-04-07  2024-04-07     1    0 days
#> 31:    2019-05-07  2025-05-25     6 2210 days
#> 32:    2015-12-02  2025-07-21     6 3519 days
#> 33:    2018-05-07  2025-01-12     2 2442 days
#> 34:    2015-12-05  2015-12-05     1    0 days
#> 35:    2016-08-29  2016-08-29     1    0 days
#> 36:    2016-07-05  2025-02-21     2 3153 days
#> 37:    2020-12-25  2022-11-20     2  695 days
#> 38:    2018-02-13  2018-02-13     1    0 days
#> 39:    2023-12-18  2023-12-18     1    0 days
#> 40:    2025-02-20  2025-02-20     1    0 days
#> 41:    2019-09-28  2019-09-28     1    0 days
#> 42:    2022-05-30  2024-12-12     3  927 days
#> 43:    2020-02-28  2020-02-28     1    0 days
#> 44:    2025-09-17  2025-09-17     1    0 days
#> 45:    2016-09-09  2016-09-09     1    0 days
#> 46:    2020-03-24  2020-03-24     1    0 days
#> 47:    2018-12-19  2018-12-19     1    0 days
#> 48:    2018-01-16  2018-01-16     1    0 days
#> 49:    2021-10-22  2021-10-22     1    0 days
#> 50:    2024-02-07  2024-02-07     1    0 days
#> 51:    2021-04-06  2021-04-06     1    0 days
#> 52:    2015-12-06  2015-12-06     1    0 days
#> 53:    2024-09-05  2024-09-05     1    0 days
#> 54:    2025-09-05  2025-09-05     1    0 days
#> 55:    2017-08-29  2017-08-29     1    0 days
#> 56:    2020-08-12  2020-08-12     1    0 days
#> 57:    2017-12-19  2017-12-19     1    0 days
#> 58:    2022-10-23  2022-10-23     1    0 days
#> 59:    2022-01-23  2022-09-27     2  247 days
#> 60:    2020-02-01  2020-02-01     1    0 days
#> 61:    2017-12-24  2017-12-24     1    0 days
#> 62:    2022-12-01  2022-12-01     1    0 days
#> 63:    2024-12-24  2024-12-24     1    0 days
#> 64:    2021-02-26  2024-07-15     2 1235 days
#> 65:    2016-12-06  2016-12-06     1    0 days
#> 66:    2024-04-04  2024-04-04     1    0 days
#> 67:    2006-02-12  2006-02-12     1    0 days
#> 68:    2011-06-18  2011-06-18     1    0 days
#> 69:    2023-10-23  2023-10-23     1    0 days
#> 70:    2022-09-15  2022-09-15     1    0 days
#> 71:    2019-10-13  2019-10-13     1    0 days
#> 72:    2016-11-21  2016-11-21     1    0 days
#> 73:    2024-01-27  2024-01-27     1    0 days
#> 74:    2017-11-19  2017-11-19     1    0 days
#> 75:    2016-01-22  2016-01-22     1    0 days
#> 76:    2014-03-23  2014-03-23     1    0 days
#> 77:    2022-02-22  2022-02-22     1    0 days
#> 78:    2020-03-10  2020-03-10     1    0 days
#> 79:    2019-01-03  2019-01-03     1    0 days
#> 80:    2020-01-12  2020-01-12     1    0 days
#> 81:    2020-10-06  2020-10-06     1    0 days
#> 82:    2021-02-13  2021-02-13     1    0 days
#> 83:    2022-10-10  2022-10-10     1    0 days
#> 84:    2014-11-15  2014-11-15     1    0 days
#> 85:    2013-07-16  2013-07-16     1    0 days
#> 86:    2015-10-29  2015-10-29     1    0 days
#> 87:    2024-05-21  2024-05-21     1    0 days
#> 88:    2005-10-09  2005-10-09     1    0 days
#> 89:    2014-06-25  2014-06-25     1    0 days
#> 90:    2021-12-03  2021-12-03     1    0 days
#> 91:    2012-07-04  2012-07-04     1    0 days
#> 92:    2012-01-25  2012-01-25     1    0 days
#> 93:    2007-05-22  2007-05-22     1    0 days
#> 94:    2021-03-18  2021-03-18     1    0 days
#> 95:    2025-09-10  2025-09-10     1    0 days
#> 96:    2010-09-27  2010-09-27     1    0 days
#> 97:    2007-01-30  2007-01-30     1    0 days
#>     firstCaseDate endCaseDate count    period
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#>