Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.

Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme.

IcdDxToCCS(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isCCSCategoryDescription = T)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD-10 using date

isCCSCategoryDescription
Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10. By default it is set to True.

Source

ICD-9-CM CCS (2015)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-CM CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip

Details

return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

IcdDxToCCS(sampleDxFile, ID, ICD, Date, "2015-10-01", TRUE)

#> Wrong ICD format: total 1 ICD codes (the number of occurrences is in brackets)

#> 123.45 (1)

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>       Short ID    ICD       Date           CCS_CATEGORY_DESCRIPTION
#>   1:   5533  C  553.3 2008-04-03                   Abdominal hernia
#>   2:   5533  A  553.3 2008-05-01                   Abdominal hernia
#>   3:  55320  A 553.20 2009-02-23                   Abdominal hernia
#>   4:  55320  B 553.20 2007-08-29                   Abdominal hernia
#>   5:   5533  A  553.3 2008-06-27                   Abdominal hernia
#>  ---                                                               
#> 166:  E03.0  B  E03.0 2014-10-01                               <NA>
#> 167:   E030  A   E030 2014-10-01 External cause codes:  Unspecified
#> 168: A01.05  B A01.05 2014-10-01                               <NA>
#> 169:   E030  C   E030 2015-10-01                  Thyroid disorders
#> 170:  A0105  A  A0105 2014-10-01                               <NA>
#> 
#> $groupedData_Long
#>     ID                                    CCS_CATEGORY_DESCRIPTION
#>  1:  C                                            Abdominal hernia
#>  2:  A                                            Abdominal hernia
#>  3:  B                                            Abdominal hernia
#>  4:  B                                            Acute bronchitis
#>  5:  A                                            Acute bronchitis
#>  6:  A                                        Cardiac dysrhythmias
#>  7:  C                                        Cardiac dysrhythmias
#>  8:  B                                        Cardiac dysrhythmias
#>  9:  A                                      Chronic kidney disease
#> 10:  C                                      Chronic kidney disease
#> 11:  B                                      Chronic kidney disease
#> 12:  C                                                       Shock
#> 13:  C                                       Heart valve disorders
#> 14:  A                                                       Shock
#> 15:  B                                                       Shock
#> 16:  C                                        Other liver diseases
#> 17:  A                                Residual codes; unclassified
#> 18:  B Other hereditary and degenerative nervous system conditions
#> 19:  C       Other nutritional; endocrine; and metabolic disorders
#> 20:  A                          External cause codes:  Unspecified
#> 21:  C                                           Thyroid disorders
#>     ID                                    CCS_CATEGORY_DESCRIPTION
#>     firstCaseDate endCaseDate count    period
#>  1:    2008-04-03  2018-12-28     2 3921 days
#>  2:    2008-05-01  2018-12-28     4 3893 days
#>  3:    2007-08-29  2019-11-10     2 4456 days
#>  4:    2008-04-29  2008-04-29     1    0 days
#>  5:    2009-06-09  2009-06-09     1    0 days
#>  6:    2007-07-27  2022-02-20    12 5322 days
#>  7:    2007-06-20  2021-09-16    19 5202 days
#>  8:    2007-07-16  2020-10-11    13 4836 days
#>  9:    2007-05-12  2022-02-08    33 5386 days
#> 10:    2007-04-21  2021-09-17    33 5263 days
#> 11:    2007-05-28  2022-03-18    34 5408 days
#> 12:    2008-03-30  2008-09-06     2  160 days
#> 13:    2007-12-12  2007-12-12     1    0 days
#> 14:    2008-05-03  2008-05-03     1    0 days
#> 15:    2008-01-19  2008-01-19     1    0 days
#> 16:    2021-01-09  2021-01-09     1    0 days
#> 17:    2020-04-29  2020-04-29     1    0 days
#> 18:    2021-06-24  2021-06-24     1    0 days
#> 19:    2019-10-16  2019-10-16     1    0 days
#> 20:    2014-10-01  2014-10-01     1    0 days
#> 21:    2015-10-01  2015-10-01     1    0 days
#>     firstCaseDate endCaseDate count    period
#> 
#> $Error
#>       ICD count IcdVersionInFile     WrongType Suggestion
#> 1: 123.45     1           ICD 10  Wrong format           
#> 2: A01.05     1            ICD 9 Wrong version           
#> 3:  E03.0     1            ICD 9 Wrong version           
#> 4:  A0105     1            ICD 9 Wrong version           
#> 5:  42761     1           ICD 10 Wrong version           
#>