Get the Multi-level Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.

Multi-leve Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme. Four levels exist in the multi-level diagnosis CCS for ICD-9-CM codes, and two levels exist in the multi-level diagnosis CCS for ICD-10-CM codes

IcdDxToCCSLvl(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, CCSLvlLabel = TRUE)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
icd 10 using date

CCSLevel
By default it is set to 1. Clinical Classifications Software (CCS) multiple level:1~4, CCS for ICD-10-CM only has 1~2 multiple levels

CCSLvlLabel
Clinical Classifications Software (CCS) multiple level categories/description for icd9/10. By default it is set to True.

Source

ICD-9-CM CCS (2012)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip

https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip

ICD-10-CM CCS (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip

Details

return Multi-leve Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

IcdDxToCCSLvl(sampleDxFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

#> Wrong ICD format: total 5 ICD codes (the number of occurrences is in brackets)

#> c("75.52 (1)", "755.2 (1)", "123.45 (1)", "001 (1)", "7552 (1)")

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDf
#>      Short ID    ICD       Date                          CCS_LVL_2_LABEL
#>   1:  5533  C  553.3 2008-04-03                         Abdominal hernia
#>   2:  5533  A  553.3 2008-05-01                         Abdominal hernia
#>   3: 55320  A 553.20 2009-02-23                         Abdominal hernia
#>   4: 55320  B 553.20 2007-08-29                         Abdominal hernia
#>   5:  5533  A  553.3 2008-06-27                         Abdominal hernia
#>  ---                                                                    
#> 171: 75.52  D  75.52 2006-05-03                                     <NA>
#> 172: 78552  A 785.52 2006-05-03 Symptoms; signs;  ill-defined conditions
#> 173:   001  B    001 2006-05-03                                     <NA>
#> 174:  7552  D   7552 2006-05-03                                     <NA>
#> 175: 755.2  D  755.2 2006-05-03                                     <NA>
#> 
#> $groupedData_Long
#>     ID                                       CCS_LVL_2_LABEL firstCaseDate
#>  1:  C                                      Abdominal hernia    2008-04-03
#>  2:  A                                      Abdominal hernia    2008-05-01
#>  3:  B                                      Abdominal hernia    2007-08-29
#>  4:  B                                Respiratory infections    2008-04-29
#>  5:  A                                Respiratory infections    2009-06-09
#>  6:  A                                 Diseases of the heart    2007-07-27
#>  7:  C                                 Diseases of the heart    2007-06-20
#>  8:  B                                 Diseases of the heart    2007-07-16
#>  9:  A                        Diseases of the urinary system    2007-05-12
#> 10:  C                        Diseases of the urinary system    2007-04-21
#> 11:  B                        Diseases of the urinary system    2007-05-28
#> 12:  C              Symptoms; signs;  ill-defined conditions    2008-03-30
#> 13:  A              Symptoms; signs;  ill-defined conditions    2006-05-03
#> 14:  B              Symptoms; signs;  ill-defined conditions    2008-01-19
#> 15:  C                                         Liver disease    2021-01-09
#> 16:  A        Residual codes; unclassified; all E codes  and    2020-04-29
#> 17:  B Hereditary and degenerative nervous system conditions    2021-06-24
#> 18:  C Other nutritional; endocrine; and metabolic disorders    2019-10-16
#> 19:  A             Residual codes; unclassified; all E codes    2014-10-01
#> 20:  C                                     Thyroid disorders    2015-10-01
#>     endCaseDate count    period
#>  1:  2018-12-28     2 3921 days
#>  2:  2018-12-28     4 3893 days
#>  3:  2019-11-10     2 4456 days
#>  4:  2008-04-29     1    0 days
#>  5:  2009-06-09     1    0 days
#>  6:  2022-02-20    12 5322 days
#>  7:  2021-09-16    20 5202 days
#>  8:  2020-10-11    13 4836 days
#>  9:  2022-02-08    33 5386 days
#> 10:  2021-09-17    33 5263 days
#> 11:  2022-03-18    34 5408 days
#> 12:  2008-09-06     2  160 days
#> 13:  2008-05-03     2  731 days
#> 14:  2008-01-19     1    0 days
#> 15:  2021-01-09     1    0 days
#> 16:  2020-04-29     1    0 days
#> 17:  2021-06-24     1    0 days
#> 18:  2019-10-16     1    0 days
#> 19:  2014-10-01     1    0 days
#> 20:  2015-10-01     1    0 days
#> 
#> $Error
#>       ICD count IcdVersionInFile     WrongType Suggestion
#> 1:  75.52     1            ICD 9  Wrong format           
#> 2: 123.45     1           ICD 10  Wrong format           
#> 3: A01.05     1            ICD 9 Wrong version           
#> 4:  E03.0     1            ICD 9 Wrong version           
#> 5:  A0105     1            ICD 9 Wrong version           
#> 6:  42761     1           ICD 10 Wrong version           
#> 7:  755.2     1            ICD 9  Wrong format     755.29
#> 8:    001     1            ICD 9  Wrong format       0019
#> 9:   7552     1            ICD 9  Wrong format      75529
#>