Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes

Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.

IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
ICD 10 using date

comorbidMethod
Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity. Change it to any of the other possible variables ('ahrq','charlson', 'elix').

Source

AHRQ

ICD-9-CM Elixhauser (2012-2015)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references

ICD-10-CM Elixhauser (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp

Charlson

ICD-9-CM Charlson (2006)

http://mchp-appserv.cpe.umanitoba.ca/Upload/SAS/ICD9_E_Charlson.sas.txt

ICD-10-CM Charlson (2006)

http://mchp-appserv.cpe.umanitoba.ca/Upload/SAS/ICD10_Charlson.sas.txt

Elixhauser

ICD-9-CM Elixhauser (2012-2015)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references

ICD-10-CM Elixhauser (2019)

https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp

Details

return comorbidity meseaures based on ICD diagnosis codes

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", charlson)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>      Short ID    ICD       Date Comorbidity
#>   1:  5533  C  553.3 2008-04-03        <NA>
#>   2:  5533  A  553.3 2008-05-01        <NA>
#>   3: 55320  A 553.20 2009-02-23        <NA>
#>   4: 55320  B 553.20 2007-08-29        <NA>
#>   5:  5533  A  553.3 2008-06-27        <NA>
#>  ---                                       
#> 271:   650  D    650 2018-10-02        <NA>
#> 272:   650  A    650 2018-10-02        <NA>
#> 273:   650  B    650 2018-10-02        <NA>
#> 274:   F42  C    F42 2018-10-02        <NA>
#> 275: Z9.90  D  Z9.90 2018-10-02        <NA>
#> 
#> $summarised_groupedDT
#>    ID Comorbidity firstCaseDate endCaseDate count    period
#> 1:  A          RD    2007-05-12  2022-02-08    33 5386 days
#> 2:  A      DIAB_C    2015-10-02  2016-10-02     2  366 days
#> 3:  B          RD    2007-05-28  2022-03-18    34 5408 days
#> 4:  B      DIAB_C    2015-10-02  2016-10-02     2  366 days
#> 5:  C          RD    2007-04-21  2021-09-17    33 5263 days
#> 6:  C      MILDLD    2021-01-09  2021-01-09     1    0 days
#> 7:  C      DIAB_C    2015-10-02  2016-10-02     2  366 days
#> 8:  D      DIAB_C    2015-10-02  2015-10-02     2    0 days
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#> 

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", elix)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>      Short ID    ICD       Date Comorbidity
#>   1:  5533  C  553.3 2008-04-03        <NA>
#>   2:  5533  A  553.3 2008-05-01        <NA>
#>   3: 55320  A 553.20 2009-02-23        <NA>
#>   4: 55320  B 553.20 2007-08-29        <NA>
#>   5:  5533  A  553.3 2008-06-27        <NA>
#>  ---                                       
#> 271:   650  D    650 2018-10-02        <NA>
#> 272:   650  A    650 2018-10-02        <NA>
#> 273:   650  B    650 2018-10-02        <NA>
#> 274:   F42  C    F42 2018-10-02        <NA>
#> 275: Z9.90  D  Z9.90 2018-10-02        <NA>
#> 
#> $summarised_groupedDT
#>     ID Comorbidity firstCaseDate endCaseDate count    period
#>  1:  A    RENLFAIL    2007-05-12  2009-04-26    28  715 days
#>  2:  A    PULMCIRC    2018-03-14  2019-08-15     3  519 days
#>  3:  B    RENLFAIL    2007-05-28  2022-03-03    29 5393 days
#>  4:  B    PULMCIRC    2019-03-11  2022-03-18     4 1103 days
#>  5:  B        METS    2021-06-24  2021-06-24     1    0 days
#>  6:  C    RENLFAIL    2007-04-21  2021-09-17    29 5263 days
#>  7:  C    PULMCIRC    2019-12-28  2020-05-10     2  134 days
#>  8:  C     HYPOTHY    2021-01-09  2021-01-09     1    0 days
#>  9:  C       NEURO    2019-10-16  2019-10-16     1    0 days
#> 10:  C    HTNWOCHF    2015-10-01  2015-10-01     1    0 days
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#> 

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", ahrq)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#> $groupedDT
#>      Short ID    ICD       Date Comorbidity
#>   1:  5533  C  553.3 2008-04-03        <NA>
#>   2:  5533  A  553.3 2008-05-01        <NA>
#>   3: 55320  A 553.20 2009-02-23        <NA>
#>   4: 55320  B 553.20 2007-08-29        <NA>
#>   5:  5533  A  553.3 2008-06-27        <NA>
#>  ---                                       
#> 271:   650  D    650 2018-10-02        <NA>
#> 272:   650  A    650 2018-10-02        <NA>
#> 273:   650  B    650 2018-10-02        <NA>
#> 274:   F42  C    F42 2018-10-02        <NA>
#> 275: Z9.90  D  Z9.90 2018-10-02        <NA>
#> 
#> $summarised_groupedDT
#>    ID Comorbidity firstCaseDate endCaseDate count    period
#> 1:  A       Renal    2007-05-12  2019-08-15    31 4478 days
#> 2:  B       Renal    2007-05-28  2022-03-18    33 5408 days
#> 3:  B  NeuroOther    2021-06-24  2021-06-24     1    0 days
#> 4:  C       Renal    2007-04-21  2021-09-17    31 5263 days
#> 5:  C       Liver    2021-01-09  2021-01-09     1    0 days
#> 6:  C     Obesity    2019-10-16  2019-10-16     1    0 days
#> 7:  C Hypothyroid    2015-10-01  2015-10-01     1    0 days
#> 
#> $Error
#>        ICD count IcdVersionInFile     WrongType Suggestion
#>  1:  A0.11    20           ICD 10  Wrong format           
#>  2:  V27.0    18           ICD 10 Wrong version           
#>  3:   E114     8           ICD 10  Wrong format           
#>  4: A01.05     8            ICD 9 Wrong version           
#>  5:  42761     7           ICD 10 Wrong version           
#>  6:  Z9.90     6           ICD 10  Wrong format           
#>  7:    F42     6           ICD 10  Wrong format           
#>  8:  V24.1     6           ICD 10 Wrong version           
#>  9:  A0105     5            ICD 9 Wrong version           
#> 10:    001     5            ICD 9  Wrong format       0019
#> 11:  75.52     4            ICD 9  Wrong format           
#> 12:  E03.0     4            ICD 9 Wrong version           
#> 13:    650     4           ICD 10 Wrong version           
#> 14: 123.45     3           ICD 10  Wrong format           
#> 15:  755.2     3            ICD 9  Wrong format     755.29
#> 16:   7552     2            ICD 9  Wrong format      75529
#>