Get the condition era

A Condition Era is defined as a span of time when the member is assumed to have a given condition. Condition Eras are periods of Condition Occurrence. Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data. Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.

getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, gapDate = 30, conditionSelect = CCS,
  isDescription = TRUE, CustomGroupingTable)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
Icd 10 using date

gapDate
Length of condition era, default is 30 days

conditionSelect
Four Stratified method can be chosen: CCS (or CCS levels), phecode (only for icd-9), comorbidities, grepICD or customICD , type `ccs`,`ccslvl1`,`ccslvl2`,`ccslvl3`,`ccslvl4`, `phecode`, `ahrq`, `charlson`, `elix` `customGrepIcdGroup`, or `customIcdGroup`, default is CCS

isDescription
CCS/Phecode categories or description for ICD-CM codes, default is True

CustomGroupingTable
Table is for groupedICDMethod:`grepICD` and `customGroup`

Details

return DxDataFile with new column, condition era.

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ccs, FALSE)

#> Wrong ICD format: total 1 ICD codes (the number of occurrences is in brackets)

#> 123.45 (1)

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID CCS_CATEGORY firstCaseDate endCaseDate count Era    period
#>  1:  C          158    2007-04-21  2021-09-17    33  16 5263 days
#>  2:  B          158    2007-05-28  2022-03-18    34  14 5408 days
#>  3:  C          106    2007-06-20  2021-09-16    19  14 5202 days
#>  4:  A          158    2007-05-12  2022-02-08    33  12 5386 days
#>  5:  B          106    2007-07-16  2020-10-11    13  10 4836 days
#>  6:  A          106    2007-07-27  2022-02-20    12   9 5322 days
#>  7:  A          143    2008-05-01  2018-12-28     4   4 3893 days
#>  8:  B          143    2007-08-29  2019-11-10     2   2 4456 days
#>  9:  C          143    2008-04-03  2018-12-28     2   2 3921 days
#> 10:  C          249    2008-03-30  2008-09-06     2   2  160 days
#> 11:  A          125    2009-06-09  2009-06-09     1   1    0 days
#> 12:  A          249    2008-05-03  2008-05-03     1   1    0 days
#> 13:  A          259    2020-04-29  2020-04-29     1   1    0 days
#> 14:  A         2620    2014-10-01  2014-10-01     1   1    0 days
#> 15:  B          125    2008-04-29  2008-04-29     1   1    0 days
#> 16:  B          249    2008-01-19  2008-01-19     1   1    0 days
#> 17:  B           81    2021-06-24  2021-06-24     1   1    0 days
#> 18:  C          151    2021-01-09  2021-01-09     1   1    0 days
#> 19:  C           48    2015-10-01  2015-10-01     1   1    0 days
#> 20:  C           58    2019-10-16  2019-10-16     1   1    0 days
#> 21:  C           96    2007-12-12  2007-12-12     1   1    0 days
#>     ID CCS_CATEGORY firstCaseDate endCaseDate count Era    period

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ICD)

#> Wrong ICD format: total 1 ICD codes (the number of occurrences is in brackets)

#> 123.45 (1)

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID    ICD firstCaseDate endCaseDate count Era    period
#>  1:  C V45.11    2007-04-21  2009-04-21    19  11  731 days
#>  2:  B V45.11    2007-05-28  2009-03-05    20  10  647 days
#>  3:  A V45.11    2007-05-30  2009-01-17    18   8  598 days
#>  4:  A  585.9    2007-09-27  2008-11-04     6   5  404 days
#>  5:  B 427.31    2007-07-16  2009-02-25     5   5  590 days
#>  6:  C I48.91    2017-06-04  2020-08-24     5   5 1177 days
#>  7:  A I48.91    2019-07-27  2021-12-13     5   4  870 days
#>  8:  C 427.31    2007-06-20  2008-09-11     5   4  449 days
#>  9:  B  585.6    2007-07-25  2008-04-28     3   3  278 days
#> 10:  B I48.91    2019-07-06  2020-10-02     3   3  454 days
#> 11:  B  N18.3    2019-03-11  2022-03-18     3   3 1103 days
#> 12:  C  585.9    2007-08-28  2009-05-26     4   3  637 days
#> 13:  A  427.1    2007-07-27  2008-12-27     2   2  519 days
#> 14:  A  553.3    2008-05-01  2008-06-27     2   2   57 days
#> 15:  B  585.9    2007-07-10  2009-03-04     2   2  603 days
#> 16:  C  585.6    2008-01-28  2009-04-06     2   2  434 days
#> 17:  C 785.52    2008-03-30  2008-09-06     2   2  160 days
#> 18:  C  N18.2    2017-11-21  2019-03-07     2   2  471 days
#> 19:  C  N18.3    2019-12-28  2020-05-10     2   2  134 days
#> 20:  A 427.31    2007-09-19  2007-09-19     1   1    0 days
#> 21:  A 427.61    2009-02-16  2009-02-16     1   1    0 days
#> 22:  A  42761    2020-03-16  2020-03-16     1   1    0 days
#> 23:  A  466.0    2009-06-09  2009-06-09     1   1    0 days
#> 24:  A 553.20    2009-02-23  2009-02-23     1   1    0 days
#> 25:  A  585.2    2007-08-08  2007-08-08     1   1    0 days
#> 26:  A  585.3    2007-08-21  2007-08-28     2   1    7 days
#> 27:  A  585.6    2009-04-26  2009-04-26     1   1    0 days
#> 28:  A  785.1    2009-03-02  2009-03-02     2   1    0 days
#> 29:  A 785.52    2008-05-03  2008-05-03     1   1    0 days
#> 30:  A  A0105    2014-10-01  2014-10-01     1   1    0 days
#> 31:  A   E030    2014-10-01  2014-10-01     1   1    0 days
#> 32:  A  I47.2    2022-02-20  2022-02-20     1   1    0 days
#> 33:  A  K44.9    2018-12-28  2018-12-28     1   1    0 days
#> 34:  A  N18.2    2022-02-08  2022-02-08     1   1    0 days
#> 35:  A  N18.3    2018-12-22  2018-12-22     1   1    0 days
#> 36:  A  N18.6    2018-03-14  2018-03-14     1   1    0 days
#> 37:  A  N18.9    2019-08-15  2019-08-15     1   1    0 days
#> 38:  A  V56.0    2007-05-12  2007-05-12     1   1    0 days
#> 39:  A  Z72.0    2020-04-29  2020-04-29     1   1    0 days
#> 40:  B  466.0    2008-04-29  2008-04-29     1   1    0 days
#> 41:  B 553.20    2007-08-29  2007-08-29     1   1    0 days
#> 42:  B  585.3    2008-07-23  2008-07-23     1   1    0 days
#> 43:  B  785.1    2008-04-07  2008-04-07     2   1    0 days
#> 44:  B 785.52    2008-01-19  2008-01-19     1   1    0 days
#> 45:  B A01.05    2014-10-01  2014-10-01     1   1    0 days
#> 46:  B  E03.0    2014-10-01  2014-10-01     1   1    0 days
#> 47:  B G25.81    2021-06-24  2021-06-24     1   1    0 days
#> 48:  B  I48.0    2019-09-01  2019-09-01     1   1    0 days
#> 49:  B  I48.2    2020-10-11  2020-10-11     1   1    0 days
#> 50:  B  I49.5    2019-11-24  2019-11-24     1   1    0 days
#> 51:  B  K44.9    2019-11-10  2019-11-10     1   1    0 days
#> 52:  B  N18.2    2018-07-17  2018-07-17     1   1    0 days
#> 53:  B  N18.9    2021-01-13  2021-01-13     1   1    0 days
#> 54:  B  V56.0    2008-09-09  2008-09-09     1   1    0 days
#> 55:  B  Z99.2    2022-02-03  2022-03-03     2   1   28 days
#> 56:  C 123.45    2020-09-20  2020-09-20     1   1    0 days
#> 57:  C  427.1    2008-03-01  2008-03-01     1   1    0 days
#> 58:  C  427.2    2008-02-16  2008-02-16     1   1    0 days
#> 59:  C 427.81    2009-02-08  2009-02-08     1   1    0 days
#> 60:  C 427.89    2008-01-23  2008-01-23     1   1    0 days
#> 61:  C  553.3    2008-04-03  2008-04-03     1   1    0 days
#> 62:  C  585.3    2007-09-26  2007-09-26     1   1    0 days
#> 63:  C  785.0    2008-12-31  2008-12-31     2   1    0 days
#> 64:  C  785.2    2007-12-12  2007-12-12     1   1    0 days
#> 65:  C   E030    2015-10-01  2015-10-01     1   1    0 days
#> 66:  C  I48.0    2021-09-16  2021-09-16     1   1    0 days
#> 67:  C  I48.2    2021-06-04  2021-06-04     1   1    0 days
#> 68:  C  K44.9    2018-12-28  2018-12-28     1   1    0 days
#> 69:  C K74.60    2021-01-09  2021-01-09     1   1    0 days
#> 70:  C  R00.0    2019-02-02  2019-02-02     1   1    0 days
#> 71:  C  V42.0    2008-08-26  2008-08-26     1   1    0 days
#> 72:  C  V56.0    2008-09-24  2008-09-24     1   1    0 days
#> 73:  C Z68.32    2019-10-16  2019-10-16     1   1    0 days
#> 74:  C  Z99.2    2021-09-17  2021-09-17     1   1    0 days
#>     ID    ICD firstCaseDate endCaseDate count Era    period

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, phecode, FALSE)

#> Wrong ICD format: total 1 ICD codes (the number of occurrences is in brackets)

#> 123.45 (1)

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID PheCode firstCaseDate endCaseDate count Era   period
#>  1:  C  585.31    2007-04-21  2009-04-21    20  11 731 days
#>  2:  B  585.31    2007-05-28  2009-03-05    21  10 647 days
#>  3:  A  585.31    2007-05-12  2009-01-17    19   8 616 days
#>  4:  A   585.3    2007-09-27  2008-11-04     6   5 404 days
#>  5:  B  427.21    2007-07-16  2009-02-25     5   5 590 days
#>  6:  C  427.21    2007-06-20  2008-09-11     5   4 449 days
#>  7:  B  585.32    2007-07-25  2008-04-28     3   3 278 days
#>  8:  C   585.3    2007-08-28  2009-05-26     4   3 637 days
#>  9:  A  427.12    2007-07-27  2008-12-27     2   2 519 days
#> 10:  A   550.2    2008-05-01  2008-06-27     2   2  57 days
#> 11:  B   585.3    2007-07-10  2009-03-04     2   2 603 days
#> 12:  C  585.32    2008-01-28  2009-04-06     2   2 434 days
#> 13:  C  994.21    2008-03-30  2008-09-06     2   2 160 days
#> 14:  A  427.21    2007-09-19  2007-09-19     1   1   0 days
#> 15:  A  427.61    2009-02-16  2009-02-16     1   1   0 days
#> 16:  A   427.9    2009-03-02  2009-03-02     2   1   0 days
#> 17:  A     483    2009-06-09  2009-06-09     1   1   0 days
#> 18:  A   550.5    2009-02-23  2009-02-23     1   1   0 days
#> 19:  A  585.32    2009-04-26  2009-04-26     1   1   0 days
#> 20:  A  585.33    2007-08-21  2007-08-28     2   1   7 days
#> 21:  A   585.4    2007-08-08  2007-08-08     1   1   0 days
#> 22:  A  994.21    2008-05-03  2008-05-03     1   1   0 days
#> 23:  B   427.9    2008-04-07  2008-04-07     2   1   0 days
#> 24:  B     483    2008-04-29  2008-04-29     1   1   0 days
#> 25:  B   550.5    2007-08-29  2007-08-29     1   1   0 days
#> 26:  B  585.33    2008-07-23  2008-07-23     1   1   0 days
#> 27:  B  994.21    2008-01-19  2008-01-19     1   1   0 days
#> 28:  C     396    2007-12-12  2007-12-12     1   1   0 days
#> 29:  C   427.1    2008-02-16  2008-02-16     1   1   0 days
#> 30:  C  427.12    2008-03-01  2008-03-01     1   1   0 days
#> 31:  C   427.3    2008-01-23  2008-01-23     1   1   0 days
#> 32:  C   427.7    2008-12-31  2008-12-31     2   1   0 days
#> 33:  C   427.8    2009-02-08  2009-02-08     1   1   0 days
#> 34:  C   550.2    2008-04-03  2008-04-03     1   1   0 days
#> 35:  C  585.33    2007-09-26  2007-09-26     1   1   0 days
#> 36:  C     587    2008-08-26  2008-08-26     1   1   0 days
#>     ID PheCode firstCaseDate endCaseDate count Era   period

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, elix)

#> Wrong ICD format: total 1 ICD codes (the number of occurrences is in brackets)

#> 123.45 (1)

#> 

#> Wrong ICD version: total 4 ICD codes (the number of occurrences is in brackets)

#> c("A01.05 (1)", "E03.0 (1)", "A0105 (1)", "42761 (1)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID Comorbidity firstCaseDate endCaseDate count Era    period
#>  1:  C    RENLFAIL    2007-04-21  2021-09-17    29  12 5263 days
#>  2:  B    RENLFAIL    2007-05-28  2022-03-03    29  10 5393 days
#>  3:  A    RENLFAIL    2007-05-12  2009-04-26    28   8  715 days
#>  4:  B    PULMCIRC    2019-03-11  2022-03-18     4   4 1103 days
#>  5:  A    PULMCIRC    2018-03-14  2019-08-15     3   3  519 days
#>  6:  C    PULMCIRC    2019-12-28  2020-05-10     2   2  134 days
#>  7:  B        METS    2021-06-24  2021-06-24     1   1    0 days
#>  8:  C    HTNWOCHF    2015-10-01  2015-10-01     1   1    0 days
#>  9:  C     HYPOTHY    2021-01-09  2021-01-09     1   1    0 days
#> 10:  C       NEURO    2019-10-16  2019-10-16     1   1    0 days

grepTable <- data.frame(group = "Cardiac dysrhythmias",
                        grepIcd = "^427|^I48",
                        stringsAsFactors = FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, customGrepIcdGroup,
                CustomGroupingTable = grepTable)

#>    ID                group firstCaseDate endCaseDate count Era    period
#> 1:  C Cardiac dysrhythmias    2007-06-20  2021-09-16    16  12 5202 days
#> 2:  A Cardiac dysrhythmias    2007-07-27  2021-12-13    10   9 5253 days
#> 3:  B Cardiac dysrhythmias    2007-07-16  2020-10-11    10   9 4836 days