Get the condition era

A Condition Era is defined as a span of time when the member is assumed to have a given condition. Condition Eras are periods of Condition Occurrence. Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data. Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.

getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupDataType = ccs, CustomGroupingTable,
  isDescription = TRUE, gapDate = 30, selectedCaseFile = NULL)

Arguments

DxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"

idColName
A column for MemberID of DxDataFile

icdColName
A column for ICD of DxDataFile

dateColName
A column for Date of DxDataFile

icd10usingDate
Icd 10 using date

groupDataType
Four Stratified methods can be chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), grepICD or customICD ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it is set to "ccs".

CustomGroupingTable
Table is for groupDataType

isDescription
CCS/PheWAS categories or description for ICD-CM codes, default is 'TRUE'.

gapDate
Length of condition era,By default it is set to 30 days "30".

selectedCaseFile
Table for selectedCases. Default is 'NULL'

Details

return DxDataFile with new column, condition era.

Examples

head(sampleDxFile)

#>    ID    ICD       Date
#> 1:  C  553.3 2008-04-03
#> 2:  A  553.3 2008-05-01
#> 3:  A 553.20 2009-02-23
#> 4:  B 553.20 2007-08-29
#> 5:  A  553.3 2008-06-27
#> 6:  B  K44.9 2019-11-10

selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl3,
                                caseCondition = "Shock",
                                caseCount = 2)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01",
                groupDataType = CCSlvl3,
                selectedCaseFile = selectedCaseFile)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

#>     ID                           CCS_LVL_3_LABEL firstCaseDate endCaseDate
#>  1:  A                          Acute bronchitis    2009-06-09  2009-06-09
#>  2:  A                      Cardiac dysrhythmias    2007-07-27  2009-03-02
#>  3:  A                    Chronic kidney disease    2007-05-12  2009-04-26
#>  4:  A                      Diaphragmatic hernia    2008-05-01  2008-06-27
#>  5:  A                    Other abdominal hernia    2009-02-23  2009-02-23
#>  6:  A Residual codes; unclassified; all E codes    2014-10-01  2014-10-01
#>  7:  A                                     Shock    2006-05-03  2008-05-03
#>  8:  B                          Acute bronchitis    2008-04-29  2008-04-29
#>  9:  B                      Cardiac dysrhythmias    2007-07-16  2009-02-25
#> 10:  B                    Chronic kidney disease    2007-05-28  2009-03-05
#> 11:  B                    Other abdominal hernia    2007-08-29  2007-08-29
#> 12:  B                                     Shock    2008-01-19  2008-01-19
#> 13:  C                      Cardiac dysrhythmias    2007-06-20  2009-02-08
#> 14:  C                    Chronic kidney disease    2007-04-21  2009-05-26
#> 15:  C                      Diaphragmatic hernia    2008-04-03  2008-04-03
#> 16:  C                     Heart valve disorders    2007-12-12  2007-12-12
#> 17:  C                                     Shock    2008-03-30  2008-09-06
#>     count era   period selectedCase
#>  1:     1   1   0 days    Selected*
#>  2:     6   4 584 days    Selected*
#>  3:    29   8 715 days    Selected*
#>  4:     2   2  57 days    Selected*
#>  5:     1   1   0 days    Selected*
#>  6:     1   1   0 days    Selected*
#>  7:     2   2 731 days    Selected*
#>  8:     1   1   0 days    Selected*
#>  9:     7   5 590 days    Selected*
#> 10:    27   9 647 days    Selected*
#> 11:     1   1   0 days    Selected*
#> 12:     1   1   0 days    Selected*
#> 13:    11   6 599 days     Selected
#> 14:    28  11 766 days     Selected
#> 15:     1   1   0 days     Selected
#> 16:     1   1   0 days     Selected
#> 17:     2   2 160 days     Selected

grepTable <- data.frame(group = "Cardiac dysrhythmias",
                        grepIcd = "^427|^I48",
                        stringsAsFactors = FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01",
                groupDataType = customGrepIcdGroup,
                CustomGroupingTable = grepTable)

#>    ID                group firstCaseDate endCaseDate count era    period
#> 1:  A Cardiac dysrhythmias    2007-07-27  2021-12-13    11  10 5253 days
#> 2:  B Cardiac dysrhythmias    2007-07-16  2020-10-11    11  10 4836 days
#> 3:  C Cardiac dysrhythmias    2007-06-20  2021-09-16    18  14 5202 days
#> 4:  D Cardiac dysrhythmias    2016-10-02  2016-10-02     2   1    0 days